#43. context_render_backend_system: Split implementation into several files as CLion has performance issues with files over 5000 lines.
