+++ unit/context_render_backend_system_vulkan/kan/context/system.c
@@ -1,2452 +1,6 @@
-#include <string.h>
-
-#include <kan/api_common/alignment.h>
-#include <kan/api_common/min_max.h>
-#include <kan/container/stack_group_allocator.h>
-#include <kan/context/render_backend_system.h>
-#include <kan/context/vulkan_memory_allocator.h>
-#include <kan/memory/allocation.h>
-#include <kan/platform/application.h>
-#include <kan/threading/atomic.h>
+#include <kan/context/render_backend_implementation_interface.h>
 
 KAN_LOG_DEFINE_CATEGORY (render_backend_system_vulkan);
-
-#define SURFACE_COLOR_FORMAT VK_FORMAT_B8G8R8A8_SRGB
-#define SURFACE_COLOR_SPACE VK_COLOR_SPACE_SRGB_NONLINEAR_KHR
-
-#if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_PROFILE_MEMORY)
-struct memory_profiling_t
-{
-    kan_allocation_group_t driver_cpu_group;
-    kan_allocation_group_t driver_cpu_generic_group;
-    kan_allocation_group_t driver_cpu_internal_group;
-
-    kan_allocation_group_t gpu_group;
-    kan_allocation_group_t gpu_unmarked_group;
-    kan_allocation_group_t gpu_buffer_group;
-    kan_allocation_group_t gpu_buffer_attribute_group;
-    kan_allocation_group_t gpu_buffer_index_group;
-    kan_allocation_group_t gpu_buffer_uniform_group;
-    kan_allocation_group_t gpu_buffer_storage_group;
-    kan_allocation_group_t gpu_image_group;
-
-    VkAllocationCallbacks vulkan_allocation_callbacks;
-    VmaDeviceMemoryCallbacks vma_device_memory_callbacks;
-};
-#endif
-
-enum surface_render_state_t
-{
-    SURFACE_RENDER_STATE_RECEIVED_NO_OUTPUT = 0u,
-    SURFACE_RENDER_STATE_RECEIVED_DATA_FROM_FRAME_BUFFER,
-    SURFACE_RENDER_STATE_RECEIVED_DATA_FROM_BLIT,
-};
-
-struct surface_blit_request_t
-{
-    struct surface_blit_request_t *next;
-    struct render_backend_image_t *image;
-    struct kan_render_integer_region_t image_region;
-    struct kan_render_integer_region_t surface_region;
-};
-
-struct surface_frame_buffer_attachment_t
-{
-    struct surface_frame_buffer_attachment_t *next;
-    struct render_backend_frame_buffer_t *frame_buffer;
-};
-
-struct render_backend_surface_t
-{
-    struct render_backend_surface_t *previous;
-    struct render_backend_surface_t *next;
-    struct render_backend_system_t *system;
-
-    kan_interned_string_t tracking_name;
-    VkSurfaceKHR surface;
-
-    VkSwapchainKHR swap_chain;
-    uint32_t images_count;
-    VkImage *images;
-    VkImageView *image_views;
-
-    VkSemaphore image_available_semaphores[KAN_CONTEXT_RENDER_BACKEND_VULKAN_FRAMES_IN_FLIGHT];
-    uint32_t acquired_image_frame;
-    uint32_t acquired_image_index;
-    kan_bool_t needs_recreation;
-
-    enum surface_render_state_t render_state;
-    struct kan_atomic_int_t blit_request_lock;
-    struct surface_blit_request_t *first_blit_request;
-    struct surface_frame_buffer_attachment_t *first_frame_buffer_attachment;
-
-    uint32_t swap_chain_creation_window_width;
-    uint32_t swap_chain_creation_window_height;
-
-    kan_application_system_window_handle_t window_handle;
-    kan_application_system_window_resource_id_t resource_id;
-};
-
-/// \details We have fully separate command state with pools and buffers for every frame in flight index,
-///          because it allows us to reset full pools instead of separate buffer which should be better from performance
-///          point of view and is advised on docs.vulkan.org.
-struct render_backend_command_state_t
-{
-    VkCommandPool graphics_command_pool;
-    VkCommandBuffer primary_graphics_command_buffer;
-
-    // TODO: Secondary graphic buffers.
-
-    VkCommandPool transfer_command_pool;
-    VkCommandBuffer primary_transfer_command_buffer;
-};
-
-struct scheduled_buffer_unmap_flush_transfer_t
-{
-    struct scheduled_buffer_unmap_flush_transfer_t *next;
-    struct render_backend_buffer_t *source_buffer;
-    struct render_backend_buffer_t *target_buffer;
-    uint64_t source_offset;
-    uint64_t target_offset;
-    uint64_t size;
-};
-
-struct scheduled_buffer_unmap_flush_t
-{
-    struct scheduled_buffer_unmap_flush_t *next;
-    struct render_backend_buffer_t *buffer;
-    uint64_t offset;
-    uint64_t size;
-};
-
-struct scheduled_image_upload_t
-{
-    struct scheduled_image_upload_t *next;
-    struct render_backend_image_t *image;
-    uint64_t mip;
-    struct render_backend_buffer_t *staging_buffer;
-    uint64_t staging_buffer_offset;
-};
-
-struct scheduled_frame_buffer_create_t
-{
-    struct scheduled_frame_buffer_create_t *next;
-    struct render_backend_frame_buffer_t *frame_buffer;
-};
-
-struct scheduled_image_mip_generation_t
-{
-    struct scheduled_image_mip_generation_t *next;
-    struct render_backend_image_t *image;
-    uint64_t first;
-    uint64_t last;
-};
-
-struct scheduled_frame_buffer_destroy_t
-{
-    struct scheduled_frame_buffer_destroy_t *next;
-    struct render_backend_frame_buffer_t *frame_buffer;
-};
-
-struct scheduled_detached_frame_buffer_destroy_t
-{
-    struct scheduled_detached_frame_buffer_destroy_t *next;
-    VkFramebuffer detached_frame_buffer;
-};
-
-struct scheduled_pass_destroy_t
-{
-    struct scheduled_pass_destroy_t *next;
-    struct render_backend_pass_t *pass;
-};
-
-struct scheduled_classic_graphics_pipeline_family_destroy_t
-{
-    struct scheduled_classic_graphics_pipeline_family_destroy_t *next;
-    struct render_backend_classic_graphics_pipeline_family_t *family;
-};
-
-struct scheduled_buffer_destroy_t
-{
-    struct scheduled_buffer_destroy_t *next;
-    struct render_backend_buffer_t *buffer;
-};
-
-struct scheduled_frame_lifetime_allocator_destroy_t
-{
-    struct scheduled_frame_lifetime_allocator_destroy_t *next;
-    struct render_backend_frame_lifetime_allocator_t *frame_lifetime_allocator;
-};
-
-struct scheduled_detached_image_view_destroy_t
-{
-    struct scheduled_detached_image_view_destroy_t *next;
-    VkImageView detached_image_view;
-};
-
-struct scheduled_image_destroy_t
-{
-    struct scheduled_image_destroy_t *next;
-    struct render_backend_image_t *image;
-};
-
-struct scheduled_detached_image_destroy_t
-{
-    struct scheduled_detached_image_destroy_t *next;
-    VkImage detached_image;
-    VmaAllocation detached_allocation;
-
-#if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_PROFILE_MEMORY)
-    uint64_t gpu_size;
-    kan_allocation_group_t gpu_allocation_group;
-#endif
-};
-
-struct render_backend_schedule_state_t
-{
-    struct kan_stack_group_allocator_t item_allocator;
-
-    /// \brief Common lock for submitting scheduled operations.
-    struct kan_atomic_int_t schedule_lock;
-
-    struct scheduled_buffer_unmap_flush_transfer_t *first_scheduled_buffer_unmap_flush_transfer;
-    struct scheduled_buffer_unmap_flush_t *first_scheduled_buffer_unmap_flush;
-    struct scheduled_image_upload_t *first_scheduled_image_upload;
-    struct scheduled_frame_buffer_create_t *first_scheduled_frame_buffer_create;
-    struct scheduled_image_mip_generation_t *first_scheduled_image_mip_generation;
-
-    struct scheduled_frame_buffer_destroy_t *first_scheduled_frame_buffer_destroy;
-    struct scheduled_detached_frame_buffer_destroy_t *first_scheduled_detached_frame_buffer_destroy;
-    struct scheduled_pass_destroy_t *first_scheduled_pass_destroy;
-    struct scheduled_classic_graphics_pipeline_family_destroy_t
-        *first_scheduled_classic_graphics_pipeline_family_destroy;
-    struct scheduled_buffer_destroy_t *first_scheduled_buffer_destroy;
-    struct scheduled_frame_lifetime_allocator_destroy_t *first_scheduled_frame_lifetime_allocator_destroy;
-    struct scheduled_detached_image_view_destroy_t *first_scheduled_detached_image_view_destroy;
-    struct scheduled_image_destroy_t *first_scheduled_image_destroy;
-    struct scheduled_detached_image_destroy_t *first_scheduled_detached_image_destroy;
-};
-
-struct render_backend_frame_buffer_attachment_t
-{
-    enum kan_render_frame_buffer_attachment_type_t type;
-    union
-    {
-        struct render_backend_image_t *image;
-        struct render_backend_surface_t *surface;
-    };
-};
-
-struct render_backend_frame_buffer_t
-{
-    struct render_backend_frame_buffer_t *next;
-    struct render_backend_frame_buffer_t *previous;
-    struct render_backend_system_t *system;
-
-    VkFramebuffer instance;
-    uint64_t instance_array_size;
-
-    /// \details If there is a surface attachment, we need to create a separate frame buffer for each surface image.
-    ///          Therefore, in that case frame_buffer_array is used instead of frame_buffer field.
-    VkFramebuffer *instance_array;
-
-    /// \brief Image views for attached images. VK_NULL_HANDLE is inserted in place of surface attachment if any.
-    VkImageView *image_views;
-
-    struct render_backend_pass_t *pass;
-    uint64_t attachments_count;
-    struct render_backend_frame_buffer_attachment_t *attachments;
-    kan_interned_string_t tracking_name;
-};
-
-struct render_backend_pass_t
-{
-    struct render_backend_pass_t *next;
-    struct render_backend_pass_t *previous;
-    struct render_backend_system_t *system;
-    VkRenderPass pass;
-};
-
-struct render_backend_layout_binding_t
-{
-    enum kan_render_layout_binding_type_t type;
-    uint64_t used_stage_mask;
-};
-
-struct render_backend_descriptor_set_layout_t
-{
-    VkDescriptorSetLayout layout;
-    kan_bool_t stable_binding;
-    uint64_t bindings_count;
-    struct render_backend_layout_binding_t bindings[];
-};
-
-struct render_backend_classic_graphics_pipeline_family_t
-{
-    struct render_backend_classic_graphics_pipeline_family_t *next;
-    struct render_backend_classic_graphics_pipeline_family_t *previous;
-    struct render_backend_system_t *system;
-
-    VkPipelineLayout layout;
-    uint64_t descriptor_set_layouts_count;
-    struct render_backend_descriptor_set_layout_t **descriptor_set_layouts;
-    kan_bool_t descriptor_sets_are_zero_sequential;
-
-    enum kan_render_classic_graphics_topology_t topology;
-    kan_interned_string_t tracking_name;
-
-    uint64_t attribute_sources_count;
-    struct kan_render_attribute_source_description_t *attribute_sources;
-
-    uint64_t attributes_count;
-    struct kan_render_attribute_description_t *attributes;
-};
-
-enum render_backend_buffer_family_t
-{
-    RENDER_BACKEND_BUFFER_FAMILY_RESOURCE = 0u,
-    RENDER_BACKEND_BUFFER_FAMILY_STAGING,
-    RENDER_BACKEND_BUFFER_FAMILY_FRAME_LIFETIME_ALLOCATOR,
-};
-
-struct render_backend_buffer_t
-{
-    struct render_backend_buffer_t *next;
-    struct render_backend_buffer_t *previous;
-    struct render_backend_system_t *system;
-
-    VkBuffer buffer;
-    VmaAllocation allocation;
-    enum render_backend_buffer_family_t family;
-    enum kan_render_buffer_type_t type;
-    uint64_t full_size;
-    kan_interned_string_t tracking_name;
-
-#if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_PROFILE_MEMORY)
-    kan_allocation_group_t device_allocation_group;
-#endif
-};
-
-#define CHUNK_FREE_MARKER UINT64_MAX
-
-struct render_backend_frame_lifetime_allocator_chunk_t
-{
-    struct render_backend_frame_lifetime_allocator_chunk_t *next;
-    struct render_backend_frame_lifetime_allocator_chunk_t *previous;
-    struct render_backend_frame_lifetime_allocator_chunk_t *next_free;
-
-    uint64_t offset;
-    uint64_t size;
-    uint64_t occupied_by_frame;
-};
-
-struct render_backend_frame_lifetime_allocator_page_t
-{
-    struct render_backend_frame_lifetime_allocator_page_t *next;
-    struct render_backend_buffer_t *buffer;
-
-    struct render_backend_frame_lifetime_allocator_chunk_t *first_chunk;
-    struct render_backend_frame_lifetime_allocator_chunk_t *first_free_chunk;
-};
-
-struct render_backend_frame_lifetime_allocator_t
-{
-    struct render_backend_frame_lifetime_allocator_t *next;
-    struct render_backend_frame_lifetime_allocator_t *previous;
-    struct render_backend_system_t *system;
-
-    struct render_backend_frame_lifetime_allocator_page_t *first_page;
-    struct render_backend_frame_lifetime_allocator_page_t *last_page;
-
-    /// \brief Common lock for allocation operations.
-    struct kan_atomic_int_t allocation_lock;
-
-    enum render_backend_buffer_family_t buffer_family;
-    enum kan_render_buffer_type_t buffer_type;
-    uint64_t page_size;
-    kan_interned_string_t tracking_name;
-};
-
-struct render_backend_frame_lifetime_allocator_allocation_t
-{
-    struct render_backend_buffer_t *buffer;
-    uint64_t offset;
-};
-
-#define STAGING_BUFFER_ALLOCATION_ALIGNMENT _Alignof (float)
-
-struct image_frame_buffer_attachment_t
-{
-    struct image_frame_buffer_attachment_t *next;
-    struct render_backend_frame_buffer_t *frame_buffer;
-};
-
-struct render_backend_image_t
-{
-    struct render_backend_image_t *next;
-    struct render_backend_image_t *previous;
-    struct render_backend_system_t *system;
-
-    VkImage image;
-    VmaAllocation allocation;
-
-    struct kan_render_image_description_t description;
-
-    /// \brief Flag used by blit-to-present routine in order to avoid
-    ///        incorrectly changing layout of the same image twice.
-    kan_bool_t switched_to_transfer_source;
-
-    struct image_frame_buffer_attachment_t *first_frame_buffer_attachment;
-
-    // TODO: Attached pipeline instances (for render target resize).
-
-#if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_PROFILE_MEMORY)
-    kan_allocation_group_t device_allocation_group;
-#endif
-};
-
-struct render_backend_system_t
-{
-    kan_context_handle_t context;
-
-    VkInstance instance;
-    VkDevice device;
-    enum kan_render_device_memory_type_t device_memory_type;
-    VkPhysicalDevice physical_device;
-
-    VkQueue graphics_queue;
-    VkQueue transfer_queue;
-
-    VmaAllocator gpu_memory_allocator;
-
-    uint32_t device_graphics_queue_family_index;
-    uint32_t device_transfer_queue_family_index;
-
-    kan_bool_t frame_started;
-    uint32_t current_frame_in_flight_index;
-
-    VkSemaphore transfer_finished_semaphores[KAN_CONTEXT_RENDER_BACKEND_VULKAN_FRAMES_IN_FLIGHT];
-    VkSemaphore render_finished_semaphores[KAN_CONTEXT_RENDER_BACKEND_VULKAN_FRAMES_IN_FLIGHT];
-    VkFence in_flight_fences[KAN_CONTEXT_RENDER_BACKEND_VULKAN_FRAMES_IN_FLIGHT];
-
-    struct render_backend_command_state_t command_states[KAN_CONTEXT_RENDER_BACKEND_VULKAN_FRAMES_IN_FLIGHT];
-    struct render_backend_schedule_state_t schedule_states[KAN_CONTEXT_RENDER_BACKEND_VULKAN_FRAMES_IN_FLIGHT];
-
-    VkFormat device_depth_image_format;
-    kan_bool_t device_depth_image_has_stencil;
-
-    /// \brief Lock for safe resource management (resource creation, memory management) in multithreaded environment.
-    /// \details Surfaces are the exemption from this rule as they're always managed from application system thread.
-    ///          Everything that is done from kan_render_backend_system_next_frame is an exemption from this rule as
-    ///          strict frame ordering must prevent any calls that are simultaneous with next frame call.
-    struct kan_atomic_int_t resource_management_lock;
-
-    struct render_backend_surface_t *first_surface;
-    struct render_backend_frame_buffer_t *first_frame_buffer;
-    struct render_backend_pass_t *first_pass;
-    struct render_backend_classic_graphics_pipeline_family_t *first_classic_graphics_pipeline_family;
-    struct render_backend_buffer_t *first_buffer;
-    struct render_backend_frame_lifetime_allocator_t *first_frame_lifetime_allocator;
-    struct render_backend_image_t *first_image;
-
-    /// \details Still listed in frame lifetime allocator list above, but referenced here too for usability.
-    struct render_backend_frame_lifetime_allocator_t *staging_frame_lifetime_allocator;
-
-#if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_VALIDATION_ENABLED)
-    kan_bool_t has_validation_layer;
-    VkDebugUtilsMessengerEXT debug_messenger;
-#endif
-
-    VmaVulkanFunctions gpu_memory_allocator_functions;
-    kan_allocation_group_t main_allocation_group;
-    kan_allocation_group_t utility_allocation_group;
-    kan_allocation_group_t surface_wrapper_allocation_group;
-    kan_allocation_group_t schedule_allocation_group;
-    kan_allocation_group_t frame_buffer_wrapper_allocation_group;
-    kan_allocation_group_t pass_wrapper_allocation_group;
-    kan_allocation_group_t pipeline_family_wrapper_allocation_group;
-    kan_allocation_group_t buffer_wrapper_allocation_group;
-    kan_allocation_group_t frame_lifetime_wrapper_allocation_group;
-    kan_allocation_group_t image_wrapper_allocation_group;
-
-    struct kan_render_supported_devices_t *supported_devices;
-
-#if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_PROFILE_MEMORY)
-    struct memory_profiling_t memory_profiling;
-#endif
-
-    kan_bool_t render_enabled;
-    kan_bool_t prefer_vsync;
-
-    kan_interned_string_t application_info_name;
-    uint64_t version_major;
-    uint64_t version_minor;
-    uint64_t version_patch;
-
-    kan_interned_string_t interned_temporary_staging_buffer;
-};
-
-#if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_PROFILE_MEMORY)
-#    define VULKAN_ALLOCATION_CALLBACKS(SYSTEM) (&(SYSTEM)->memory_profiling.vulkan_allocation_callbacks)
-
-static void *profiled_allocate (void *user_data, size_t size, size_t alignment, VkSystemAllocationScope scope)
-{
-    struct memory_profiling_t *profiling = (struct memory_profiling_t *) user_data;
-    uint64_t real_alignment = KAN_MAX (alignment, _Alignof (uint64_t));
-    uint64_t real_size = kan_apply_alignment (real_alignment + size, real_alignment);
-
-    void *allocated_data = kan_allocate_general (profiling->driver_cpu_generic_group, real_size, real_alignment);
-
-    void *accessible_data = ((uint8_t *) allocated_data) + real_alignment;
-    uint64_t *meta_output = allocated_data;
-    *meta_output = real_size;
-    ++meta_output;
-
-    while ((uint8_t *) meta_output < (uint8_t *) accessible_data)
-    {
-        *meta_output = 0u;
-        ++meta_output;
-    }
-
-    return accessible_data;
-}
-
-static inline void *walk_from_accessible_to_allocated (void *accessible)
-{
-    uint64_t *meta = accessible;
-    do
-    {
-        meta--;
-    } while (*meta == 0u);
-
-    return meta;
-}
-
-static inline void profiled_free_internal (struct memory_profiling_t *profiling, void *allocated, void *accessible)
-{
-    uint64_t size_to_free = (*(uint64_t *) allocated) + (((uint8_t *) accessible) - ((uint8_t *) allocated));
-    kan_free_general (profiling->driver_cpu_generic_group, allocated, size_to_free);
-}
-
-static void *profiled_reallocate (
-    void *user_data, void *original, size_t size, size_t alignment, VkSystemAllocationScope scope)
-{
-    struct memory_profiling_t *profiling = (struct memory_profiling_t *) user_data;
-    void *original_user_accessible_data = original;
-    void *original_allocated_data = walk_from_accessible_to_allocated (original_user_accessible_data);
-
-    const uint64_t original_real_size = *(uint64_t *) original_allocated_data;
-    const uint64_t original_data_size =
-        original_real_size - (((uint8_t *) original_user_accessible_data) - ((uint8_t *) original_allocated_data));
-
-    void *new_data = profiled_allocate (user_data, size, alignment, scope);
-    memcpy (new_data, original, KAN_MIN ((size_t) original_data_size, size));
-    profiled_free_internal (profiling, original_allocated_data, original_user_accessible_data);
-    return new_data;
-}
-
-static void profiled_free (void *user_data, void *pointer)
-{
-    if (pointer)
-    {
-        struct memory_profiling_t *profiling = (struct memory_profiling_t *) user_data;
-        void *accessible = pointer;
-        void *allocated = walk_from_accessible_to_allocated (accessible);
-        profiled_free_internal (profiling, allocated, accessible);
-    }
-}
-
-static void notify_internal_cpu_allocation (void *user_data,
-                                            size_t size,
-                                            VkInternalAllocationType type,
-                                            VkSystemAllocationScope scope)
-{
-    struct memory_profiling_t *profiling = (struct memory_profiling_t *) user_data;
-    kan_allocation_group_allocate (profiling->driver_cpu_internal_group, size);
-}
-
-static void notify_internal_cpu_free (void *user_data,
-                                      size_t size,
-                                      VkInternalAllocationType type,
-                                      VkSystemAllocationScope scope)
-{
-    struct memory_profiling_t *profiling = (struct memory_profiling_t *) user_data;
-    kan_allocation_group_free (profiling->driver_cpu_internal_group, size);
-}
-
-static void notify_device_allocation (
-    VmaAllocator allocator, uint32_t memory_type, VkDeviceMemory memory, VkDeviceSize size, void *user_data)
-{
-    struct memory_profiling_t *profiling = (struct memory_profiling_t *) user_data;
-    kan_allocation_group_allocate (profiling->gpu_unmarked_group, size);
-}
-
-static void notify_device_free (
-    VmaAllocator allocator, uint32_t memory_type, VkDeviceMemory memory, VkDeviceSize size, void *user_data)
-{
-    struct memory_profiling_t *profiling = (struct memory_profiling_t *) user_data;
-    kan_allocation_group_free (profiling->gpu_unmarked_group, size);
-}
-
-static void memory_profiling_init (struct render_backend_system_t *system)
-{
-    system->memory_profiling.driver_cpu_group =
-        kan_allocation_group_get_child (system->main_allocation_group, "driver_cpu");
-    system->memory_profiling.driver_cpu_generic_group =
-        kan_allocation_group_get_child (system->memory_profiling.driver_cpu_group, "generic");
-    system->memory_profiling.driver_cpu_internal_group =
-        kan_allocation_group_get_child (system->memory_profiling.driver_cpu_group, "internal");
-
-    system->memory_profiling.gpu_group = kan_allocation_group_get_child (system->main_allocation_group, "gpu");
-    system->memory_profiling.gpu_unmarked_group =
-        kan_allocation_group_get_child (system->memory_profiling.gpu_group, "unmarked");
-    system->memory_profiling.gpu_buffer_group =
-        kan_allocation_group_get_child (system->memory_profiling.gpu_group, "buffer");
-    system->memory_profiling.gpu_buffer_attribute_group =
-        kan_allocation_group_get_child (system->memory_profiling.gpu_buffer_group, "attribute");
-    system->memory_profiling.gpu_buffer_index_group =
-        kan_allocation_group_get_child (system->memory_profiling.gpu_buffer_group, "index");
-    system->memory_profiling.gpu_buffer_uniform_group =
-        kan_allocation_group_get_child (system->memory_profiling.gpu_buffer_group, "uniform");
-    system->memory_profiling.gpu_buffer_storage_group =
-        kan_allocation_group_get_child (system->memory_profiling.gpu_buffer_group, "storage");
-    system->memory_profiling.gpu_image_group =
-        kan_allocation_group_get_child (system->memory_profiling.gpu_group, "image");
-
-    system->memory_profiling.vulkan_allocation_callbacks = (VkAllocationCallbacks) {
-        .pUserData = &system->memory_profiling,
-        .pfnAllocation = profiled_allocate,
-        .pfnReallocation = profiled_reallocate,
-        .pfnFree = profiled_free,
-        .pfnInternalAllocation = notify_internal_cpu_allocation,
-        .pfnInternalFree = notify_internal_cpu_free,
-    };
-
-    system->memory_profiling.vma_device_memory_callbacks = (VmaDeviceMemoryCallbacks) {
-        .pfnAllocate = notify_device_allocation,
-        .pfnFree = notify_device_free,
-        .pUserData = &system->memory_profiling,
-    };
-}
-
-static inline void transfer_memory_between_groups (uint64_t amount,
-                                                   kan_allocation_group_t from,
-                                                   kan_allocation_group_t to)
-{
-    kan_allocation_group_free (from, amount);
-    kan_allocation_group_allocate (to, amount);
-}
-#else
-#    define VULKAN_ALLOCATION_CALLBACKS(SYSTEM) NULL
-#endif
-
-static inline struct render_backend_schedule_state_t *render_backend_system_get_schedule_for_memory (
-    struct render_backend_system_t *system)
-{
-    // Always schedule into the current frame, even if it is not yet started.
-    return &system->schedule_states[system->current_frame_in_flight_index];
-}
-
-static inline struct render_backend_schedule_state_t *render_backend_system_get_schedule_for_destroy (
-    struct render_backend_system_t *system)
-{
-    uint64_t schedule_index = system->current_frame_in_flight_index;
-    if (!system->frame_started)
-    {
-        // If frame is not started, then we can't schedule destroy for current frame as destroy is done when frame
-        // starts and therefore we have a risk of destroying used resources. Technically, if frame is noy yet started,
-        // then resource to be destroyed will not be used in it, therefore the most safe and logical solution is to
-        // add destroy into previous frame schedule.
-        schedule_index =
-            schedule_index == 0u ? KAN_CONTEXT_RENDER_BACKEND_VULKAN_FRAMES_IN_FLIGHT - 1u : schedule_index - 1u;
-    }
-
-    return &system->schedule_states[schedule_index];
-}
-
-static inline struct render_backend_frame_buffer_t *render_backend_system_create_frame_buffer (
-    struct render_backend_system_t *system, struct kan_render_frame_buffer_description_t *description)
-{
-    struct render_backend_frame_buffer_t *buffer = kan_allocate_batched (system->frame_buffer_wrapper_allocation_group,
-                                                                         sizeof (struct render_backend_frame_buffer_t));
-
-    buffer->next = system->first_frame_buffer;
-    buffer->previous = NULL;
-    buffer->system = system;
-
-    if (system->first_frame_buffer)
-    {
-        system->first_frame_buffer->previous = buffer;
-    }
-
-    system->first_frame_buffer = buffer;
-    buffer->instance = VK_NULL_HANDLE;
-    buffer->instance_array_size = 0u;
-    buffer->instance_array = NULL;
-    buffer->image_views = NULL;
-
-    buffer->pass = (struct render_backend_pass_t *) description->associated_pass;
-    buffer->tracking_name = description->tracking_name;
-
-    buffer->attachments_count = description->attachment_count;
-    buffer->attachments =
-        kan_allocate_general (system->frame_buffer_wrapper_allocation_group,
-                              sizeof (struct render_backend_frame_buffer_attachment_t) * buffer->attachments_count,
-                              _Alignof (struct render_backend_frame_buffer_attachment_t));
-
-    for (uint64_t index = 0u; index < buffer->attachments_count; ++index)
-    {
-        struct kan_render_frame_buffer_attachment_description_t *source = &description->attachments[index];
-        struct render_backend_frame_buffer_attachment_t *target = &buffer->attachments[index];
-        target->type = source->type;
-
-        switch (source->type)
-        {
-        case KAN_FRAME_BUFFER_ATTACHMENT_IMAGE:
-        {
-            target->image = (struct render_backend_image_t *) source->image;
-            struct image_frame_buffer_attachment_t *attachment = kan_allocate_batched (
-                system->image_wrapper_allocation_group, sizeof (struct image_frame_buffer_attachment_t));
-
-            attachment->next = target->image->first_frame_buffer_attachment;
-            target->image->first_frame_buffer_attachment = attachment;
-            attachment->frame_buffer = buffer;
-            break;
-        }
-
-        case KAN_FRAME_BUFFER_ATTACHMENT_SURFACE:
-        {
-            target->surface = (struct render_backend_surface_t *) source->surface;
-            struct surface_frame_buffer_attachment_t *attachment = kan_allocate_batched (
-                system->surface_wrapper_allocation_group, sizeof (struct surface_frame_buffer_attachment_t));
-
-            attachment->next = target->surface->first_frame_buffer_attachment;
-            target->surface->first_frame_buffer_attachment = attachment;
-            attachment->frame_buffer = buffer;
-            break;
-        }
-        }
-    }
-
-    struct render_backend_schedule_state_t *schedule = render_backend_system_get_schedule_for_memory (system);
-    kan_atomic_int_lock (&schedule->schedule_lock);
-
-    struct scheduled_frame_buffer_create_t *item =
-        KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (&schedule->item_allocator, struct scheduled_frame_buffer_create_t);
-
-    // We do not need to preserve order as frame buffers cannot depend one on another.
-    item->next = schedule->first_scheduled_frame_buffer_create;
-    schedule->first_scheduled_frame_buffer_create = item;
-    item->frame_buffer = buffer;
-    kan_atomic_int_unlock (&schedule->schedule_lock);
-    return buffer;
-}
-
-static inline void render_backend_frame_buffer_destroy_resources (struct render_backend_system_t *system,
-                                                                  struct render_backend_frame_buffer_t *frame_buffer)
-{
-    if (frame_buffer->instance != VK_NULL_HANDLE)
-    {
-        vkDestroyFramebuffer (system->device, frame_buffer->instance, VULKAN_ALLOCATION_CALLBACKS (system));
-        frame_buffer->instance = VK_NULL_HANDLE;
-    }
-
-    if (frame_buffer->instance_array)
-    {
-        for (uint64_t index = 0u; index < frame_buffer->instance_array_size; ++index)
-        {
-            vkDestroyFramebuffer (system->device, frame_buffer->instance_array[index],
-                                  VULKAN_ALLOCATION_CALLBACKS (system));
-        }
-
-        kan_free_general (system->frame_buffer_wrapper_allocation_group, frame_buffer->instance_array,
-                          sizeof (VkFramebuffer) * frame_buffer->instance_array_size);
-        frame_buffer->instance_array_size = 0u;
-        frame_buffer->instance_array = NULL;
-    }
-
-    if (frame_buffer->image_views)
-    {
-        for (uint64_t index = 0u; index < frame_buffer->attachments_count; ++index)
-        {
-            if (frame_buffer->image_views[index] != VK_NULL_HANDLE)
-            {
-                vkDestroyImageView (system->device, frame_buffer->image_views[index],
-                                    VULKAN_ALLOCATION_CALLBACKS (system));
-            }
-        }
-
-        kan_free_general (system->frame_buffer_wrapper_allocation_group, frame_buffer->image_views,
-                          sizeof (VkImageView) * frame_buffer->attachments_count);
-        frame_buffer->image_views = NULL;
-    }
-}
-
-static inline void render_backend_frame_buffer_schedule_resource_destroy (
-    struct render_backend_system_t *system,
-    struct render_backend_frame_buffer_t *frame_buffer,
-    struct render_backend_schedule_state_t *schedule)
-{
-    if (frame_buffer->instance != VK_NULL_HANDLE)
-    {
-        struct scheduled_detached_frame_buffer_destroy_t *frame_buffer_destroy =
-            KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (&schedule->item_allocator,
-                                                      struct scheduled_detached_frame_buffer_destroy_t);
-        frame_buffer_destroy->next = schedule->first_scheduled_detached_frame_buffer_destroy;
-        schedule->first_scheduled_detached_frame_buffer_destroy = frame_buffer_destroy;
-        frame_buffer_destroy->detached_frame_buffer = frame_buffer->instance;
-        frame_buffer->instance = VK_NULL_HANDLE;
-    }
-
-    if (frame_buffer->instance_array)
-    {
-        for (uint64_t index = 0u; index < frame_buffer->instance_array_size; ++index)
-        {
-            struct scheduled_detached_frame_buffer_destroy_t *frame_buffer_destroy =
-                KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (&schedule->item_allocator,
-                                                          struct scheduled_detached_frame_buffer_destroy_t);
-            frame_buffer_destroy->next = schedule->first_scheduled_detached_frame_buffer_destroy;
-            schedule->first_scheduled_detached_frame_buffer_destroy = frame_buffer_destroy;
-            frame_buffer_destroy->detached_frame_buffer = frame_buffer->instance_array[index];
-        }
-
-        kan_free_general (system->frame_buffer_wrapper_allocation_group, frame_buffer->instance_array,
-                          sizeof (VkFramebuffer) * frame_buffer->instance_array_size);
-        frame_buffer->instance_array_size = 0u;
-        frame_buffer->instance_array = NULL;
-    }
-
-    if (frame_buffer->image_views)
-    {
-        for (uint64_t index = 0u; index < frame_buffer->attachments_count; ++index)
-        {
-            if (frame_buffer->image_views[index] != VK_NULL_HANDLE)
-            {
-                struct scheduled_detached_image_view_destroy_t *image_view_destroy =
-                    KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (&schedule->item_allocator,
-                                                              struct scheduled_detached_image_view_destroy_t);
-                image_view_destroy->next = schedule->first_scheduled_detached_image_view_destroy;
-                schedule->first_scheduled_detached_image_view_destroy = image_view_destroy;
-                image_view_destroy->detached_image_view = frame_buffer->image_views[index];
-            }
-        }
-
-        kan_free_general (system->frame_buffer_wrapper_allocation_group, frame_buffer->image_views,
-                          sizeof (VkImageView) * frame_buffer->attachments_count);
-        frame_buffer->image_views = NULL;
-    }
-}
-
-static inline void render_backend_system_destroy_frame_buffer (struct render_backend_system_t *system,
-                                                               struct render_backend_frame_buffer_t *frame_buffer,
-                                                               kan_bool_t remove_from_list)
-{
-    if (remove_from_list)
-    {
-        if (frame_buffer->next)
-        {
-            frame_buffer->next->previous = frame_buffer->previous;
-        }
-
-        if (frame_buffer->previous)
-        {
-            frame_buffer->previous->next = frame_buffer->next;
-        }
-        else
-        {
-            KAN_ASSERT (system->first_frame_buffer = frame_buffer)
-            system->first_frame_buffer = frame_buffer->next;
-        }
-    }
-
-    // Detach from surface and images if attachment is created.
-    for (uint64_t index = 0u; index < frame_buffer->attachments_count; ++index)
-    {
-        struct render_backend_frame_buffer_attachment_t *attachment = &frame_buffer->attachments[index];
-        switch (attachment->type)
-        {
-        case KAN_FRAME_BUFFER_ATTACHMENT_IMAGE:
-        {
-            struct image_frame_buffer_attachment_t *previous = NULL;
-            struct image_frame_buffer_attachment_t *current = attachment->image->first_frame_buffer_attachment;
-
-            while (current)
-            {
-                if (current->frame_buffer == frame_buffer)
-                {
-                    if (previous)
-                    {
-                        previous->next = current->next;
-                    }
-                    else
-                    {
-                        KAN_ASSERT (attachment->image->first_frame_buffer_attachment == current)
-                        attachment->image->first_frame_buffer_attachment = current->next;
-                    }
-
-                    kan_free_batched (system->image_wrapper_allocation_group, current);
-                    break;
-                }
-
-                previous = current;
-                current = current->next;
-            }
-
-            break;
-        }
-
-        case KAN_FRAME_BUFFER_ATTACHMENT_SURFACE:
-        {
-            struct surface_frame_buffer_attachment_t *previous = NULL;
-            struct surface_frame_buffer_attachment_t *current = attachment->surface->first_frame_buffer_attachment;
-
-            while (current)
-            {
-                if (current->frame_buffer == frame_buffer)
-                {
-                    if (previous)
-                    {
-                        previous->next = current->next;
-                    }
-                    else
-                    {
-                        KAN_ASSERT (attachment->surface->first_frame_buffer_attachment == current)
-                        attachment->surface->first_frame_buffer_attachment = current->next;
-                    }
-
-                    kan_free_batched (system->surface_wrapper_allocation_group, current);
-                    break;
-                }
-
-                previous = current;
-                current = current->next;
-            }
-
-            break;
-        }
-        }
-    }
-
-    render_backend_frame_buffer_destroy_resources (system, frame_buffer);
-    kan_free_general (system->frame_buffer_wrapper_allocation_group, frame_buffer->attachments,
-                      sizeof (struct render_backend_frame_buffer_attachment_t) * frame_buffer->attachments_count);
-    kan_free_batched (system->frame_buffer_wrapper_allocation_group, frame_buffer);
-}
-
-static inline struct render_backend_pass_t *render_backend_system_create_pass (
-    struct render_backend_system_t *system, struct kan_render_pass_description_t *description)
-{
-    VkAttachmentDescription *attachment_descriptions = kan_allocate_general (
-        system->utility_allocation_group, sizeof (VkAttachmentDescription) * description->attachments_count,
-        _Alignof (VkAttachmentDescription));
-
-    VkAttachmentReference *color_attachments = kan_allocate_general (
-        system->utility_allocation_group, sizeof (VkAttachmentReference) * description->attachments_count,
-        _Alignof (VkAttachmentReference));
-
-    VkAttachmentReference *next_color_attachment = color_attachments;
-    uint64_t color_attachments_count = 0u;
-    kan_bool_t has_depth_attachment = KAN_FALSE;
-    VkAttachmentReference depth_attachment;
-
-    for (uint64_t index = 0u; index < description->attachments_count; ++index)
-    {
-        struct kan_render_pass_attachment_t *attachment = &description->attachments[index];
-        struct VkAttachmentDescription *vulkan_attachment = &attachment_descriptions[index];
-        vulkan_attachment->flags = 0u;
-
-        switch (attachment->type)
-        {
-        case KAN_RENDER_PASS_ATTACHMENT_COLOR:
-            ++color_attachments_count;
-
-            switch (attachment->color_format)
-            {
-            case KAN_RENDER_COLOR_FORMAT_RGBA32_SRGB:
-                vulkan_attachment->format = VK_FORMAT_R8G8B8A8_SRGB;
-                break;
-
-            case KAN_RENDER_COLOR_FORMAT_RGBA128_SFLOAT:
-                vulkan_attachment->format = VK_FORMAT_R32G32B32A32_SFLOAT;
-                break;
-
-            case KAN_RENDER_COLOR_FORMAT_SURFACE:
-                vulkan_attachment->format = SURFACE_COLOR_FORMAT;
-                break;
-            }
-
-            switch (attachment->samples)
-            {
-            case 1u:
-                vulkan_attachment->samples = VK_SAMPLE_COUNT_1_BIT;
-                break;
-
-            case 2u:
-                vulkan_attachment->samples = VK_SAMPLE_COUNT_2_BIT;
-                break;
-
-            case 4u:
-                vulkan_attachment->samples = VK_SAMPLE_COUNT_4_BIT;
-                break;
-
-            case 8u:
-                vulkan_attachment->samples = VK_SAMPLE_COUNT_8_BIT;
-                break;
-
-            case 16u:
-                vulkan_attachment->samples = VK_SAMPLE_COUNT_16_BIT;
-                break;
-
-            case 32u:
-                vulkan_attachment->samples = VK_SAMPLE_COUNT_32_BIT;
-                break;
-
-            case 64u:
-                vulkan_attachment->samples = VK_SAMPLE_COUNT_64_BIT;
-                break;
-
-            default:
-                // Count of samples is not power of two.
-                KAN_ASSERT (KAN_FALSE)
-                break;
-            }
-
-            vulkan_attachment->finalLayout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-            next_color_attachment->attachment = (uint32_t) index;
-            next_color_attachment->layout = VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL;
-            ++next_color_attachment;
-            break;
-
-        case KAN_RENDER_PASS_ATTACHMENT_DEPTH_STENCIL:
-            // There should be not more than 1 depth attachment per pass.
-            KAN_ASSERT (!has_depth_attachment)
-
-            has_depth_attachment = KAN_TRUE;
-            vulkan_attachment->format = system->device_depth_image_format;
-            KAN_ASSERT (attachment->samples == 1u)
-            vulkan_attachment->samples = VK_SAMPLE_COUNT_1_BIT;
-            vulkan_attachment->finalLayout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-
-            depth_attachment.attachment = (uint32_t) index;
-            depth_attachment.layout = VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
-            break;
-        }
-
-        switch (attachment->load_operation)
-        {
-        case KAN_RENDER_LOAD_OPERATION_ANY:
-            vulkan_attachment->loadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-            break;
-
-        case KAN_RENDER_LOAD_OPERATION_LOAD:
-            vulkan_attachment->loadOp = VK_ATTACHMENT_LOAD_OP_LOAD;
-            break;
-
-        case KAN_RENDER_LOAD_OPERATION_CLEAR:
-            vulkan_attachment->loadOp = VK_ATTACHMENT_LOAD_OP_CLEAR;
-            break;
-        }
-
-        switch (attachment->store_operation)
-        {
-        case KAN_RENDER_STORE_OPERATION_ANY:
-            vulkan_attachment->storeOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-            break;
-
-        case KAN_RENDER_STORE_OPERATION_STORE:
-            vulkan_attachment->storeOp = VK_ATTACHMENT_STORE_OP_STORE;
-            break;
-
-        case KAN_RENDER_STORE_OPERATION_NONE:
-            vulkan_attachment->storeOp = VK_ATTACHMENT_STORE_OP_NONE;
-            break;
-        }
-
-        vulkan_attachment->stencilLoadOp = VK_ATTACHMENT_LOAD_OP_DONT_CARE;
-        vulkan_attachment->stencilStoreOp = VK_ATTACHMENT_STORE_OP_DONT_CARE;
-        vulkan_attachment->initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    }
-
-    VkSubpassDescription sub_pass_description = {
-        .flags = 0u,
-        .pipelineBindPoint = VK_PIPELINE_BIND_POINT_GRAPHICS,
-        .inputAttachmentCount = 0u,
-        .pInputAttachments = NULL,
-        .colorAttachmentCount = (uint32_t) color_attachments_count,
-        .pColorAttachments = color_attachments,
-        .pResolveAttachments = NULL,
-        .pDepthStencilAttachment = has_depth_attachment ? &depth_attachment : NULL,
-        .preserveAttachmentCount = 0u,
-        .pPreserveAttachments = NULL,
-    };
-
-    uint32_t destination_access_mask = 0u;
-    if (color_attachments_count > 0u)
-    {
-        destination_access_mask |= VK_ACCESS_COLOR_ATTACHMENT_WRITE_BIT;
-    }
-
-    if (has_depth_attachment)
-    {
-        destination_access_mask |= VK_ACCESS_DEPTH_STENCIL_ATTACHMENT_WRITE_BIT;
-    }
-
-    VkSubpassDependency sub_pass_dependencies[] = {
-        {
-            .srcSubpass = VK_SUBPASS_EXTERNAL,
-            .dstSubpass = 0u,
-            .srcStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT | VK_PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT,
-            .dstStageMask = VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT | VK_PIPELINE_STAGE_EARLY_FRAGMENT_TESTS_BIT,
-            .srcAccessMask = 0u,
-            .dstAccessMask = destination_access_mask,
-            .dependencyFlags = VK_DEPENDENCY_BY_REGION_BIT,
-        },
-    };
-
-    VkRenderPassCreateInfo render_pass_description = {
-        .sType = VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO,
-        .pNext = NULL,
-        .flags = 0u,
-        .attachmentCount = (uint32_t) description->attachments_count,
-        .pAttachments = attachment_descriptions,
-        .subpassCount = 1u,
-        .pSubpasses = &sub_pass_description,
-        .dependencyCount = 1u,
-        .pDependencies = sub_pass_dependencies,
-    };
-
-    VkRenderPass render_pass;
-    VkResult result = vkCreateRenderPass (system->device, &render_pass_description,
-                                          VULKAN_ALLOCATION_CALLBACKS (system), &render_pass);
-
-    kan_free_general (system->utility_allocation_group, attachment_descriptions,
-                      sizeof (VkAttachmentDescription) * description->attachments_count);
-    kan_free_general (system->utility_allocation_group, color_attachments,
-                      sizeof (VkAttachmentReference) * description->attachments_count);
-
-    if (result != VK_SUCCESS)
-    {
-        KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR, "Failed to create render pass \"%s\".",
-                 description->tracking_name)
-        return NULL;
-    }
-
-    struct render_backend_pass_t *pass =
-        kan_allocate_batched (system->pass_wrapper_allocation_group, sizeof (struct render_backend_pass_t));
-
-    pass->next = system->first_pass;
-    pass->previous = NULL;
-    pass->system = system;
-
-    if (system->first_pass)
-    {
-        system->first_pass->previous = pass;
-    }
-
-    system->first_pass = pass;
-    pass->pass = render_pass;
-    return pass;
-}
-
-static inline void free_descriptor_set_layouts (struct render_backend_system_t *system,
-                                                uint64_t descriptor_set_layouts_count,
-                                                struct render_backend_descriptor_set_layout_t **descriptor_set_layouts)
-{
-    for (uint64_t index = 0u; index < descriptor_set_layouts_count; ++index)
-    {
-        if (descriptor_set_layouts[index] != NULL)
-        {
-            if (descriptor_set_layouts[index]->layout != VK_NULL_HANDLE)
-            {
-                vkDestroyDescriptorSetLayout (system->device, descriptor_set_layouts[index]->layout,
-                                              VULKAN_ALLOCATION_CALLBACKS (system));
-            }
-
-            kan_free_general (
-                system->pipeline_family_wrapper_allocation_group, descriptor_set_layouts[index],
-                sizeof (struct render_backend_descriptor_set_layout_t) +
-                    sizeof (struct render_backend_layout_binding_t) * descriptor_set_layouts[index]->bindings_count);
-        }
-    }
-
-    kan_free_general (system->pipeline_family_wrapper_allocation_group, descriptor_set_layouts,
-                      sizeof (struct render_backend_descriptor_set_layout_t *) * descriptor_set_layouts_count);
-}
-
-static inline struct render_backend_classic_graphics_pipeline_family_t *
-render_backend_system_create_classic_graphics_pipeline_family (
-    struct render_backend_system_t *system,
-    struct kan_render_classic_graphics_pipeline_family_description_t *description)
-{
-    kan_bool_t descriptor_sets_are_zero_sequential = KAN_TRUE;
-    uint64_t sets_count = 0u;
-
-    for (uint64_t index = 0u; index < description->layouts_count; ++index)
-    {
-        if (description->layouts[index].set != index)
-        {
-            KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR,
-                     "Pipeline family \"%s\" layout set indices do not form zero-based sequence with step equal to "
-                     "one. It makes bind less efficient.",
-                     description->tracking_name)
-            descriptor_sets_are_zero_sequential = KAN_FALSE;
-            break;
-        }
-
-        sets_count = KAN_MAX (sets_count, description->layouts[index].set + 1u);
-    }
-
-    if (sets_count >= KAN_CONTEXT_RENDER_BACKEND_VULKAN_MAX_SET_INDEX)
-    {
-        KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR,
-                 "Pipeline family \"%s\" last layout set index %lu is higher than maximum allowed %lu.",
-                 description->tracking_name, (unsigned long) (sets_count - 1u),
-                 (unsigned long) KAN_CONTEXT_RENDER_BACKEND_VULKAN_MAX_SET_INDEX)
-        return NULL;
-    }
-
-    VkPipelineLayout pipeline_layout;
-    struct render_backend_descriptor_set_layout_t **descriptor_set_layouts =
-        kan_allocate_general (system->pipeline_family_wrapper_allocation_group,
-                              sizeof (struct render_backend_descriptor_set_layout_t *) * sets_count,
-                              _Alignof (struct render_backend_descriptor_set_layout_t *));
-
-    for (uint64_t index = 0u; index < sets_count; ++index)
-    {
-        descriptor_set_layouts[index] = NULL;
-    }
-
-    kan_bool_t descriptor_set_layouts_created = KAN_TRUE;
-    uint64_t bindings_size = 0u;
-    VkDescriptorSetLayoutBinding *bindings = NULL;
-
-    VkDescriptorSetLayout *descriptor_set_layouts_for_pipeline = kan_allocate_general (
-        system->utility_allocation_group, sizeof (VkDescriptorSetLayout) * description->layouts_count,
-        _Alignof (VkDescriptorSetLayout));
-
-    for (uint64_t layout_index = 0u; layout_index < description->layouts_count; ++layout_index)
-    {
-        struct kan_render_layout_description_t *layout_description = &description->layouts[layout_index];
-        if (descriptor_set_layouts[layout_description->set])
-        {
-            KAN_LOG (
-                render_backend_system_vulkan, KAN_LOG_ERROR,
-                "Failed to create descriptor set for layout %lu for pipeline family \"%s\" as set %lu is already used.",
-                (unsigned long) layout_index, description->tracking_name, (unsigned long) layout_description->set)
-            descriptor_set_layouts_created = KAN_FALSE;
-            break;
-        }
-
-        if (bindings_size < layout_description->bindings_count)
-        {
-            if (bindings)
-            {
-                kan_free_general (system->utility_allocation_group, bindings,
-                                  sizeof (VkDescriptorSetLayoutBinding) * bindings_size);
-            }
-
-            bindings_size = layout_description->bindings_count;
-            bindings = kan_allocate_general (system->utility_allocation_group,
-                                             sizeof (VkDescriptorSetLayoutBinding) * bindings_size,
-                                             _Alignof (VkDescriptorSetLayoutBinding));
-        }
-
-        uint64_t bindings_count = 0u;
-        for (uint64_t binding_index = 0u; binding_index < layout_description->bindings_count; ++binding_index)
-        {
-            struct kan_render_layout_binding_description_t *binding_description =
-                &layout_description->bindings[binding_index];
-            VkDescriptorSetLayoutBinding *vulkan_binding = &bindings[binding_index];
-
-            vulkan_binding->binding = (uint32_t) binding_description->binding;
-            vulkan_binding->descriptorCount = 1u;
-            vulkan_binding->stageFlags = 0u;
-            vulkan_binding->pImmutableSamplers = NULL;
-
-            switch (binding_description->type)
-            {
-            case KAN_RENDER_LAYOUT_BINDING_TYPE_UNIFORM_BUFFER:
-                vulkan_binding->descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-                break;
-
-            case KAN_RENDER_LAYOUT_BINDING_TYPE_STORAGE_BUFFER:
-                vulkan_binding->descriptorType = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-                break;
-
-            case KAN_RENDER_LAYOUT_BINDING_TYPE_COMBINED_IMAGE_SAMPLER:
-                vulkan_binding->descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-                break;
-            }
-
-            if (binding_description->used_stage_mask & (1u << KAN_RENDER_STAGE_CLASSIC_GRAPHICS_VERTEX))
-            {
-                vulkan_binding->stageFlags |= VK_SHADER_STAGE_VERTEX_BIT;
-            }
-
-            if (binding_description->used_stage_mask & (1u << KAN_RENDER_STAGE_CLASSIC_GRAPHICS_FRAGMENT))
-            {
-                vulkan_binding->stageFlags |= VK_SHADER_STAGE_FRAGMENT_BIT;
-            }
-
-            bindings_count = KAN_MAX (bindings_count, binding_description->binding + 1u);
-        }
-
-        if (bindings_count >= KAN_CONTEXT_RENDER_BACKEND_VULKAN_MAX_BINDING_INDEX)
-        {
-            KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR,
-                     "Pipeline family \"%s\" descriptor set layout %lu last binding index %lu is higher than maximum "
-                     "allowed %lu.",
-                     description->tracking_name, (unsigned long) layout_index, (unsigned long) (sets_count - 1u),
-                     (unsigned long) KAN_CONTEXT_RENDER_BACKEND_VULKAN_MAX_BINDING_INDEX)
-            descriptor_set_layouts_created = KAN_FALSE;
-            break;
-        }
-
-        VkDescriptorSetLayoutCreateInfo layout_info = {
-            .sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO,
-            .pNext = NULL,
-            .flags = 0u,
-            .bindingCount = (uint32_t) layout_description->bindings_count,
-            .pBindings = bindings,
-        };
-
-        if (vkCreateDescriptorSetLayout (system->device, &layout_info, VULKAN_ALLOCATION_CALLBACKS (system),
-                                         &descriptor_set_layouts_for_pipeline[layout_index]) != VK_SUCCESS)
-        {
-            KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR,
-                     "Failed to create descriptor set for layout %lu for pipeline family \"%s\".",
-                     (unsigned long) layout_index, description->tracking_name)
-            descriptor_set_layouts_created = KAN_FALSE;
-            break;
-        }
-
-        struct render_backend_descriptor_set_layout_t *layout =
-            kan_allocate_general (system->pipeline_family_wrapper_allocation_group,
-                                  sizeof (struct render_backend_descriptor_set_layout_t) +
-                                      sizeof (struct render_backend_layout_binding_t) * bindings_count,
-                                  _Alignof (struct render_backend_descriptor_set_layout_t));
-
-        layout->layout = descriptor_set_layouts_for_pipeline[layout_index];
-        layout->stable_binding = layout_description->stable_binding;
-        layout->bindings_count = bindings_count;
-
-        for (uint64_t binding = 0u; binding < bindings_count; ++binding)
-        {
-            layout->bindings[binding].type = KAN_RENDER_LAYOUT_BINDING_TYPE_UNIFORM_BUFFER;
-            // Bindings with zero used stage mask are treated as non-existent.
-            layout->bindings[binding].used_stage_mask = 0u;
-        }
-
-        for (uint64_t binding_index = 0u; binding_index < layout_description->bindings_count; ++binding_index)
-        {
-            struct kan_render_layout_binding_description_t *binding_description =
-                &layout_description->bindings[binding_index];
-            layout->bindings[binding_description->binding].type = binding_description->type;
-            layout->bindings[binding_description->binding].used_stage_mask = binding_description->used_stage_mask;
-        }
-    }
-
-    if (bindings)
-    {
-        kan_free_general (system->utility_allocation_group, bindings,
-                          sizeof (VkDescriptorSetLayoutBinding) * bindings_size);
-    }
-
-    if (!descriptor_set_layouts_created)
-    {
-        free_descriptor_set_layouts (system, sets_count, descriptor_set_layouts);
-        kan_free_general (system->utility_allocation_group, descriptor_set_layouts_for_pipeline,
-                          sizeof (VkDescriptorSetLayout) * description->layouts_count);
-        return NULL;
-    }
-
-    VkPipelineLayoutCreateInfo pipeline_layout_info = {
-        .sType = VK_STRUCTURE_TYPE_PIPELINE_LAYOUT_CREATE_INFO,
-        .pNext = NULL,
-        .flags = 0u,
-        .setLayoutCount = (uint32_t) description->layouts_count,
-        .pSetLayouts = descriptor_set_layouts_for_pipeline,
-        .pushConstantRangeCount = 0u,
-        .pPushConstantRanges = NULL,
-    };
-
-    VkResult result = vkCreatePipelineLayout (system->device, &pipeline_layout_info,
-                                              VULKAN_ALLOCATION_CALLBACKS (system), &pipeline_layout);
-    kan_free_general (system->utility_allocation_group, descriptor_set_layouts_for_pipeline,
-                      sizeof (VkDescriptorSetLayout) * description->layouts_count);
-
-    if (result != VK_SUCCESS)
-    {
-        KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR, "Failed to pipeline layout for pipeline family \"%s\".",
-                 description->tracking_name)
-        free_descriptor_set_layouts (system, sets_count, descriptor_set_layouts);
-        return NULL;
-    }
-
-    struct render_backend_classic_graphics_pipeline_family_t *family =
-        kan_allocate_batched (system->pipeline_family_wrapper_allocation_group,
-                              sizeof (struct render_backend_classic_graphics_pipeline_family_t));
-
-    family->next = system->first_classic_graphics_pipeline_family;
-    family->previous = NULL;
-    family->system = system;
-
-    if (system->first_classic_graphics_pipeline_family)
-    {
-        system->first_classic_graphics_pipeline_family->previous = family;
-    }
-
-    system->first_classic_graphics_pipeline_family = family;
-    family->layout = pipeline_layout;
-    family->descriptor_set_layouts_count = sets_count;
-    family->descriptor_set_layouts = descriptor_set_layouts;
-    family->descriptor_sets_are_zero_sequential = descriptor_sets_are_zero_sequential;
-
-    family->topology = description->topology;
-    family->tracking_name = description->tracking_name;
-
-#define COPY_POD_ARRAY(TO, FROM, NAME, TYPE)                                                                           \
-    (TO).NAME##_count = (FROM).NAME##_count;                                                                           \
-    if ((TO).NAME##_count > 0u)                                                                                        \
-    {                                                                                                                  \
-        (TO).NAME = kan_allocate_general (system->pipeline_family_wrapper_allocation_group,                            \
-                                          sizeof (TYPE) * (TO).NAME##_count, _Alignof (TYPE));                         \
-        memcpy ((TO).NAME, (FROM).NAME, sizeof (TYPE) * (TO).NAME##_count);                                            \
-    }                                                                                                                  \
-    else                                                                                                               \
-    {                                                                                                                  \
-        (TO).NAME = NULL;                                                                                              \
-    }
-
-    COPY_POD_ARRAY (*family, *description, attribute_sources, struct kan_render_attribute_source_description_t)
-    COPY_POD_ARRAY (*family, *description, attributes, struct kan_render_attribute_description_t)
-#undef COPY_POD_ARRAY
-
-    return family;
-}
-
-static inline void render_backend_system_destroy_classic_graphics_pipeline_family (
-    struct render_backend_system_t *system,
-    struct render_backend_classic_graphics_pipeline_family_t *family,
-    kan_bool_t remove_from_list)
-{
-    if (remove_from_list)
-    {
-        if (family->next)
-        {
-            family->next->previous = family->previous;
-        }
-
-        if (family->previous)
-        {
-            family->previous->next = family->next;
-        }
-        else
-        {
-            KAN_ASSERT (system->first_classic_graphics_pipeline_family = family)
-            system->first_classic_graphics_pipeline_family = family->next;
-        }
-    }
-
-    vkDestroyPipelineLayout (system->device, family->layout, VULKAN_ALLOCATION_CALLBACKS (system));
-    free_descriptor_set_layouts (system, family->descriptor_set_layouts_count, family->descriptor_set_layouts);
-
-    kan_free_general (system->pipeline_family_wrapper_allocation_group, family->attribute_sources,
-                      sizeof (struct kan_render_attribute_source_description_t) * family->attribute_sources_count);
-    kan_free_general (system->pipeline_family_wrapper_allocation_group, family->attributes,
-                      sizeof (struct kan_render_attribute_description_t) * family->attributes_count);
-}
-
-static inline void render_backend_system_destroy_pass (struct render_backend_system_t *system,
-                                                       struct render_backend_pass_t *pass,
-                                                       kan_bool_t remove_from_list)
-{
-    if (remove_from_list)
-    {
-        if (pass->next)
-        {
-            pass->next->previous = pass->previous;
-        }
-
-        if (pass->previous)
-        {
-            pass->previous->next = pass->next;
-        }
-        else
-        {
-            KAN_ASSERT (system->first_pass = pass)
-            system->first_pass = pass->next;
-        }
-    }
-
-    vkDestroyRenderPass (system->device, pass->pass, VULKAN_ALLOCATION_CALLBACKS (system));
-    kan_free_batched (system->pass_wrapper_allocation_group, pass);
-}
-
-static inline struct render_backend_buffer_t *render_backend_system_create_buffer (
-    struct render_backend_system_t *system,
-    enum render_backend_buffer_family_t family,
-    enum kan_render_buffer_type_t buffer_type,
-    uint64_t full_size,
-    kan_interned_string_t tracking_name)
-{
-    VkBufferUsageFlags usage_flags = 0u;
-    VmaAllocationCreateFlagBits allocation_flags = 0u;
-
-    switch (system->device_memory_type)
-    {
-    case KAN_RENDER_DEVICE_MEMORY_TYPE_SEPARATE:
-        break;
-
-    case KAN_RENDER_DEVICE_MEMORY_TYPE_UNIFIED:
-    case KAN_RENDER_DEVICE_MEMORY_TYPE_UNIFIED_COHERENT:
-        // All memory is unified and host visible anyway.
-        allocation_flags = VMA_ALLOCATION_CREATE_HOST_ACCESS_SEQUENTIAL_WRITE_BIT;
-        break;
-    }
-
-    switch (family)
-    {
-    case RENDER_BACKEND_BUFFER_FAMILY_RESOURCE:
-        usage_flags |= VK_BUFFER_USAGE_TRANSFER_DST_BIT;
-        break;
-
-    case RENDER_BACKEND_BUFFER_FAMILY_STAGING:
-        usage_flags |= VK_BUFFER_USAGE_TRANSFER_SRC_BIT;
-        allocation_flags = VMA_ALLOCATION_CREATE_HOST_ACCESS_SEQUENTIAL_WRITE_BIT;
-        break;
-
-    case RENDER_BACKEND_BUFFER_FAMILY_FRAME_LIFETIME_ALLOCATOR:
-        allocation_flags = VMA_ALLOCATION_CREATE_HOST_ACCESS_SEQUENTIAL_WRITE_BIT;
-        break;
-    }
-
-    if (family != RENDER_BACKEND_BUFFER_FAMILY_STAGING)
-    {
-        switch (buffer_type)
-        {
-        case KAN_RENDER_BUFFER_TYPE_ATTRIBUTE:
-            usage_flags |= VK_BUFFER_USAGE_VERTEX_BUFFER_BIT;
-            break;
-
-        case KAN_RENDER_BUFFER_TYPE_INDEX_16:
-        case KAN_RENDER_BUFFER_TYPE_INDEX_32:
-            usage_flags |= VK_BUFFER_USAGE_INDEX_BUFFER_BIT;
-            break;
-
-        case KAN_RENDER_BUFFER_TYPE_UNIFORM:
-            usage_flags |= VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT;
-            break;
-
-        case KAN_RENDER_BUFFER_TYPE_STORAGE:
-            usage_flags |= VK_BUFFER_USAGE_STORAGE_BUFFER_BIT;
-            break;
-        }
-    }
-
-    uint32_t used_queues_count;
-    uint32_t used_queues_indices[2u];
-
-    switch (family)
-    {
-    case RENDER_BACKEND_BUFFER_FAMILY_RESOURCE:
-        if (system->device_graphics_queue_family_index == system->device_transfer_queue_family_index)
-        {
-            used_queues_count = 1u;
-            used_queues_indices[0u] = system->device_graphics_queue_family_index;
-        }
-        else
-        {
-            used_queues_count = 2u;
-            used_queues_indices[0u] = system->device_graphics_queue_family_index;
-            used_queues_indices[0u] = system->device_transfer_queue_family_index;
-        }
-
-        break;
-
-    case RENDER_BACKEND_BUFFER_FAMILY_STAGING:
-        used_queues_count = 1u;
-        used_queues_indices[0u] = system->device_transfer_queue_family_index;
-        break;
-
-    case RENDER_BACKEND_BUFFER_FAMILY_FRAME_LIFETIME_ALLOCATOR:
-        used_queues_count = 1u;
-        used_queues_indices[0u] = system->device_graphics_queue_family_index;
-        break;
-    }
-
-    VkBufferCreateInfo buffer_create_info = {
-        .sType = VK_STRUCTURE_TYPE_BUFFER_CREATE_INFO,
-        .pNext = NULL,
-        .flags = 0u,
-        .size = full_size,
-        .usage = usage_flags,
-        .sharingMode = used_queues_count <= 1u ? VK_SHARING_MODE_EXCLUSIVE : VK_SHARING_MODE_CONCURRENT,
-        .queueFamilyIndexCount = used_queues_count,
-        .pQueueFamilyIndices = used_queues_indices,
-    };
-
-    VmaAllocationCreateInfo allocation_create_info = {
-        .flags = allocation_flags,
-        .usage = VMA_MEMORY_USAGE_AUTO,
-        .requiredFlags = 0u,
-        .preferredFlags = 0u,
-        .memoryTypeBits = UINT32_MAX,
-        .pool = NULL,
-        .pUserData = NULL,
-        .priority = 0.0f,
-    };
-
-    VmaAllocationInfo allocation_info;
-    VkBuffer buffer_handle;
-    VmaAllocation allocation_handle;
-
-    if (vmaCreateBuffer (system->gpu_memory_allocator, &buffer_create_info, &allocation_create_info, &buffer_handle,
-                         &allocation_handle, &allocation_info) != VK_SUCCESS)
-    {
-        KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR, "Failed to create buffer \"%s\" of size %llu.",
-                 tracking_name, (unsigned long long) full_size)
-        return NULL;
-    }
-
-    struct render_backend_buffer_t *buffer =
-        kan_allocate_batched (system->buffer_wrapper_allocation_group, sizeof (struct render_backend_buffer_t));
-
-    buffer->next = system->first_buffer;
-    buffer->previous = NULL;
-
-    if (system->first_buffer)
-    {
-        system->first_buffer->previous = buffer;
-    }
-
-    system->first_buffer = buffer;
-    buffer->system = system;
-    buffer->buffer = buffer_handle;
-    buffer->allocation = allocation_handle;
-    buffer->family = family;
-    buffer->type = buffer_type;
-    buffer->full_size = full_size;
-    buffer->tracking_name = tracking_name;
-
-#if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_PROFILE_MEMORY)
-    kan_allocation_group_t type_allocation_group = system->memory_profiling.gpu_buffer_group;
-    switch (buffer_type)
-    {
-    case KAN_RENDER_BUFFER_TYPE_ATTRIBUTE:
-        type_allocation_group = system->memory_profiling.gpu_buffer_attribute_group;
-        break;
-
-    case KAN_RENDER_BUFFER_TYPE_INDEX_16:
-    case KAN_RENDER_BUFFER_TYPE_INDEX_32:
-        type_allocation_group = system->memory_profiling.gpu_buffer_index_group;
-        break;
-
-    case KAN_RENDER_BUFFER_TYPE_UNIFORM:
-        type_allocation_group = system->memory_profiling.gpu_buffer_uniform_group;
-        break;
-
-    case KAN_RENDER_BUFFER_TYPE_STORAGE:
-        type_allocation_group = system->memory_profiling.gpu_buffer_storage_group;
-        break;
-    }
-
-    buffer->device_allocation_group = kan_allocation_group_get_child (type_allocation_group, buffer->tracking_name);
-    transfer_memory_between_groups (full_size, system->memory_profiling.gpu_unmarked_group,
-                                    buffer->device_allocation_group);
-#endif
-
-    return buffer;
-}
-
-static inline void render_backend_system_destroy_buffer (struct render_backend_system_t *system,
-                                                         struct render_backend_buffer_t *buffer,
-                                                         kan_bool_t remove_from_list)
-{
-    if (remove_from_list)
-    {
-        if (buffer->next)
-        {
-            buffer->next->previous = buffer->previous;
-        }
-
-        if (buffer->previous)
-        {
-            buffer->previous->next = buffer->next;
-        }
-        else
-        {
-            KAN_ASSERT (system->first_buffer = buffer)
-            system->first_buffer = buffer->next;
-        }
-    }
-
-#if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_PROFILE_MEMORY)
-    transfer_memory_between_groups (buffer->full_size, buffer->device_allocation_group,
-                                    system->memory_profiling.gpu_unmarked_group);
-#endif
-
-    vmaDestroyBuffer (system->gpu_memory_allocator, buffer->buffer, buffer->allocation);
-    kan_free_batched (system->buffer_wrapper_allocation_group, buffer);
-}
-
-static struct render_backend_frame_lifetime_allocator_t *render_backend_system_create_frame_lifetime_allocator (
-    struct render_backend_system_t *system,
-    enum render_backend_buffer_family_t buffer_family,
-    enum kan_render_buffer_type_t buffer_type,
-    uint64_t page_size,
-    kan_interned_string_t tracking_name)
-{
-    struct render_backend_frame_lifetime_allocator_t *allocator = kan_allocate_batched (
-        system->frame_lifetime_wrapper_allocation_group, sizeof (struct render_backend_frame_lifetime_allocator_t));
-
-    allocator->next = system->first_frame_lifetime_allocator;
-    allocator->previous = NULL;
-
-    if (system->first_frame_lifetime_allocator)
-    {
-        system->first_frame_lifetime_allocator->previous = allocator;
-    }
-
-    system->first_frame_lifetime_allocator = allocator;
-    allocator->system = system;
-
-    allocator->first_page = NULL;
-    allocator->last_page = NULL;
-    allocator->allocation_lock = kan_atomic_int_init (0);
-
-    KAN_ASSERT (buffer_family == RENDER_BACKEND_BUFFER_FAMILY_STAGING ||
-                buffer_family == RENDER_BACKEND_BUFFER_FAMILY_FRAME_LIFETIME_ALLOCATOR)
-    allocator->buffer_family = buffer_family;
-    allocator->buffer_type = buffer_type;
-    allocator->page_size = page_size;
-    allocator->tracking_name = tracking_name;
-    return allocator;
-}
-
-static struct render_backend_frame_lifetime_allocator_allocation_t
-render_backend_frame_lifetime_allocator_allocate_on_page (struct render_backend_frame_lifetime_allocator_t *allocator,
-                                                          struct render_backend_frame_lifetime_allocator_page_t *page,
-                                                          uint64_t size,
-                                                          uint64_t alignment)
-{
-    struct render_backend_frame_lifetime_allocator_allocation_t result = {
-        .buffer = NULL,
-        .offset = 0u,
-    };
-
-    struct render_backend_frame_lifetime_allocator_chunk_t *previous_chunk = NULL;
-    struct render_backend_frame_lifetime_allocator_chunk_t *chunk = page->first_free_chunk;
-
-    while (chunk)
-    {
-        KAN_ASSERT (chunk->occupied_by_frame == CHUNK_FREE_MARKER)
-        const uint64_t memory_offset = chunk->offset;
-        const uint64_t allocation_offset = kan_apply_alignment (memory_offset, alignment);
-        const uint64_t allocation_size = size + allocation_offset - memory_offset;
-
-        if (chunk->size >= allocation_size)
-        {
-            // Can allocate from here.
-            result.buffer = page->buffer;
-            result.offset = allocation_offset;
-
-            chunk->offset += allocation_size;
-            chunk->size -= allocation_size;
-
-            if (chunk->previous &&
-                chunk->previous->occupied_by_frame == allocator->system->current_frame_in_flight_index)
-            {
-                // Previous chunk is associated with current frame, increase it by allocation.
-                chunk->previous->size += allocation_size;
-
-                // If current chunk is fully consumed by this operation, delete it.
-                if (chunk->size == 0u)
-                {
-                    chunk->previous->next = chunk->next;
-                    if (chunk->next)
-                    {
-                        chunk->next->previous = chunk->previous;
-                    }
-
-                    if (previous_chunk)
-                    {
-                        previous_chunk->next_free = chunk->next_free;
-                    }
-                    else
-                    {
-                        KAN_ASSERT (page->first_free_chunk)
-                        page->first_free_chunk = chunk->next_free;
-                    }
-
-                    kan_free_batched (allocator->system->frame_lifetime_wrapper_allocation_group, chunk);
-                }
-            }
-            else if (chunk->size == 0u)
-            {
-                // Chunk was fully used by this allocation. Undo it and mark chunk as occupied by current frame.
-                chunk->offset = memory_offset;
-                chunk->size = allocation_size;
-                chunk->occupied_by_frame = allocator->system->current_frame_in_flight_index;
-
-                // Remove from free list.
-                if (previous_chunk)
-                {
-                    previous_chunk->next_free = chunk->next_free;
-                }
-                else
-                {
-                    KAN_ASSERT (page->first_free_chunk)
-                    page->first_free_chunk = chunk->next_free;
-                }
-            }
-            else
-            {
-                // We need to allocate new chunk in order to represent this allocation.
-                struct render_backend_frame_lifetime_allocator_chunk_t *new_chunk =
-                    kan_allocate_batched (allocator->system->frame_lifetime_wrapper_allocation_group,
-                                          sizeof (struct render_backend_frame_lifetime_allocator_chunk_t));
-
-                new_chunk->offset = memory_offset;
-                new_chunk->size = allocation_size;
-                new_chunk->occupied_by_frame = allocator->system->current_frame_in_flight_index;
-
-                new_chunk->next_free = NULL;
-                new_chunk->next = chunk;
-                new_chunk->previous = chunk->previous;
-                chunk->previous = new_chunk;
-
-                if (new_chunk->previous)
-                {
-                    new_chunk->previous->next = new_chunk;
-                }
-                else
-                {
-                    KAN_ASSERT (page->first_chunk == chunk)
-                    page->first_chunk = new_chunk;
-                }
-            }
-
-            return result;
-        }
-
-        previous_chunk = chunk;
-        chunk = chunk->next_free;
-    }
-
-    return result;
-}
-
-static struct render_backend_frame_lifetime_allocator_allocation_t render_backend_frame_lifetime_allocator_allocate (
-    struct render_backend_frame_lifetime_allocator_t *allocator, uint64_t size, uint64_t alignment)
-{
-    struct render_backend_frame_lifetime_allocator_allocation_t result = {
-        .buffer = NULL,
-        .offset = 0u,
-    };
-
-    if (size > allocator->page_size)
-    {
-        KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR,
-                 "Frame-lifetime allocator \"%s\": Caught attempt to allocate data of size %llu which is greater than "
-                 "page size %llu.",
-                 allocator->tracking_name, (unsigned long long) size, (unsigned long long) allocator->page_size)
-        return result;
-    }
-
-    kan_atomic_int_lock (&allocator->allocation_lock);
-    struct render_backend_frame_lifetime_allocator_page_t *page = allocator->first_page;
-
-    while (page)
-    {
-        result = render_backend_frame_lifetime_allocator_allocate_on_page (allocator, page, size, alignment);
-        if (result.buffer)
-        {
-            kan_atomic_int_unlock (&allocator->allocation_lock);
-            return result;
-        }
-
-        page = page->next;
-    }
-
-    // No space on existing pages, need to create new one.
-    struct render_backend_buffer_t *new_page_buffer =
-        render_backend_system_create_buffer (allocator->system, allocator->buffer_family, allocator->buffer_type,
-                                             allocator->page_size, allocator->tracking_name);
-
-    if (!new_page_buffer)
-    {
-        KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR,
-                 "Frame-lifetime allocator \"%s\": Failed to allocate %llu bytes as all pages are full and new page "
-                 "allocation has failed.",
-                 allocator->tracking_name, (unsigned long long) size)
-
-        kan_atomic_int_unlock (&allocator->allocation_lock);
-        return result;
-    }
-
-    struct render_backend_frame_lifetime_allocator_page_t *new_page =
-        kan_allocate_batched (allocator->system->frame_lifetime_wrapper_allocation_group,
-                              sizeof (struct render_backend_frame_lifetime_allocator_page_t));
-
-    new_page->next = NULL;
-    if (allocator->last_page)
-    {
-        allocator->last_page->next = new_page;
-    }
-    else
-    {
-        allocator->first_page = new_page;
-    }
-
-    allocator->last_page = new_page;
-    new_page->buffer = new_page_buffer;
-
-    new_page->first_chunk = kan_allocate_batched (allocator->system->frame_lifetime_wrapper_allocation_group,
-                                                  sizeof (struct render_backend_frame_lifetime_allocator_chunk_t));
-    new_page->first_free_chunk = new_page->first_chunk;
-
-    new_page->first_chunk->next = NULL;
-    new_page->first_chunk->previous = NULL;
-    new_page->first_chunk->next_free = NULL;
-    new_page->first_chunk->offset = 0u;
-    new_page->first_chunk->size = allocator->page_size;
-    new_page->first_chunk->occupied_by_frame = CHUNK_FREE_MARKER;
-
-    result = render_backend_frame_lifetime_allocator_allocate_on_page (allocator, new_page, size, alignment);
-    kan_atomic_int_unlock (&allocator->allocation_lock);
-    return result;
-}
-
-static struct render_backend_frame_lifetime_allocator_allocation_t render_backend_system_allocate_for_staging (
-    struct render_backend_system_t *system, uint64_t size)
-{
-    if (size <= system->staging_frame_lifetime_allocator->page_size)
-    {
-        // Size is not overwhelmingly big, so we can use staging frame lifetime allocator.
-        return render_backend_frame_lifetime_allocator_allocate (system->staging_frame_lifetime_allocator, size,
-                                                                 STAGING_BUFFER_ALLOCATION_ALIGNMENT);
-    }
-
-    // Size is super big, we need separate buffer.
-    kan_atomic_int_lock (&system->resource_management_lock);
-    struct render_backend_buffer_t *buffer = render_backend_system_create_buffer (
-        system, RENDER_BACKEND_BUFFER_FAMILY_STAGING, KAN_RENDER_BUFFER_TYPE_STORAGE, size,
-        system->interned_temporary_staging_buffer);
-    kan_atomic_int_unlock (&system->resource_management_lock);
-
-    if (!buffer)
-    {
-        KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR,
-                 "Failed to stage %llu bytes: allocation is too big for staging allocator and there is no memory for "
-                 "separate staging buffer.",
-                 (unsigned long long) size)
-
-        return (struct render_backend_frame_lifetime_allocator_allocation_t) {
-            .buffer = NULL,
-            .offset = 0u,
-        };
-    }
-
-    // Schedule deletion of temporary buffer.
-    struct render_backend_schedule_state_t *schedule = render_backend_system_get_schedule_for_destroy (system);
-    kan_atomic_int_lock (&schedule->schedule_lock);
-
-    struct scheduled_buffer_destroy_t *item =
-        KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (&schedule->item_allocator, struct scheduled_buffer_destroy_t);
-
-    // We do not need to preserve order as buffers cannot depend one on another.
-    item->next = schedule->first_scheduled_buffer_destroy;
-    schedule->first_scheduled_buffer_destroy = item;
-    item->buffer = buffer;
-    kan_atomic_int_unlock (&schedule->schedule_lock);
-
-    return (struct render_backend_frame_lifetime_allocator_allocation_t) {
-        .buffer = buffer,
-        .offset = 0u,
-    };
-}
-
-static void render_backend_frame_lifetime_allocator_retire_old_allocations (
-    struct render_backend_frame_lifetime_allocator_t *allocator)
-{
-    struct render_backend_frame_lifetime_allocator_page_t *page = allocator->first_page;
-    while (page)
-    {
-        struct render_backend_frame_lifetime_allocator_chunk_t *chunk = page->first_chunk;
-        while (chunk)
-        {
-            struct render_backend_frame_lifetime_allocator_chunk_t *next_chunk = chunk->next;
-            if (chunk->occupied_by_frame == allocator->system->current_frame_in_flight_index)
-            {
-                // Chunk should retire as its data lifetime has ended.
-                chunk->occupied_by_frame = CHUNK_FREE_MARKER;
-            }
-
-            if (chunk->occupied_by_frame == CHUNK_FREE_MARKER && chunk->previous &&
-                chunk->previous->occupied_by_frame == CHUNK_FREE_MARKER)
-            {
-                // We can merge this chunk into previous free chunk.
-                chunk->previous->size += chunk->size;
-                chunk->previous->next = chunk->next;
-
-                if (chunk->next)
-                {
-                    chunk->next->previous = chunk->previous;
-                }
-
-                kan_free_batched (allocator->system->frame_lifetime_wrapper_allocation_group, chunk);
-            }
-
-            chunk = next_chunk;
-        }
-
-        page->first_free_chunk = NULL;
-        chunk = page->first_chunk;
-
-        while (chunk)
-        {
-            if (chunk->occupied_by_frame == CHUNK_FREE_MARKER)
-            {
-                chunk->next_free = page->first_free_chunk;
-                page->first_free_chunk = chunk;
-            }
-
-            chunk = chunk->next;
-        }
-
-        page = page->next;
-    }
-}
-
-static void render_backend_frame_lifetime_allocator_destroy_page (
-    struct render_backend_system_t *system,
-    struct render_backend_frame_lifetime_allocator_page_t *page,
-    kan_bool_t destroy_buffers)
-{
-    if (destroy_buffers)
-    {
-        render_backend_system_destroy_buffer (system, page->buffer, KAN_TRUE);
-    }
-
-    struct render_backend_frame_lifetime_allocator_chunk_t *chunk = page->first_chunk;
-    while (chunk)
-    {
-        struct render_backend_frame_lifetime_allocator_chunk_t *next_chunk = chunk->next;
-        kan_free_batched (system->frame_lifetime_wrapper_allocation_group, chunk);
-        chunk = next_chunk;
-    }
-
-    kan_free_batched (system->frame_lifetime_wrapper_allocation_group, page);
-}
-
-static void render_backend_frame_lifetime_allocator_clean_empty_pages (
-    struct render_backend_frame_lifetime_allocator_t *allocator)
-{
-    struct render_backend_frame_lifetime_allocator_page_t *previous_page = NULL;
-    struct render_backend_frame_lifetime_allocator_page_t *page = allocator->first_page;
-
-    while (page)
-    {
-        struct render_backend_frame_lifetime_allocator_page_t *next_page = page->next;
-        if (page->first_free_chunk && page->first_free_chunk->offset == 0u &&
-            page->first_free_chunk->size == allocator->page_size)
-        {
-            KAN_ASSERT (!page->first_free_chunk->next_free)
-            // Looks like page is empty. Lets get rid of it.
-
-            if (previous_page)
-            {
-                previous_page->next = next_page;
-            }
-            else
-            {
-                KAN_ASSERT (page == allocator->first_page)
-                allocator->first_page = page->next;
-            }
-
-            if (!page->next)
-            {
-                allocator->last_page = previous_page;
-            }
-
-            render_backend_frame_lifetime_allocator_destroy_page (allocator->system, page, KAN_TRUE);
-        }
-        else
-        {
-            // Page is not empty, we should keep it.
-            previous_page = page;
-        }
-
-        page = next_page;
-    }
-}
-
-static void render_backend_system_destroy_frame_lifetime_allocator (
-    struct render_backend_system_t *system,
-    struct render_backend_frame_lifetime_allocator_t *frame_lifetime_allocator,
-    kan_bool_t destroy_buffers,
-    kan_bool_t remove_from_list)
-{
-    if (remove_from_list)
-    {
-        if (frame_lifetime_allocator->next)
-        {
-            frame_lifetime_allocator->next->previous = frame_lifetime_allocator->previous;
-        }
-
-        if (frame_lifetime_allocator->previous)
-        {
-            frame_lifetime_allocator->previous->next = frame_lifetime_allocator->next;
-        }
-        else
-        {
-            KAN_ASSERT (system->first_frame_lifetime_allocator = frame_lifetime_allocator)
-            system->first_frame_lifetime_allocator = frame_lifetime_allocator->next;
-        }
-    }
-
-    struct render_backend_frame_lifetime_allocator_page_t *page = frame_lifetime_allocator->first_page;
-    while (page)
-    {
-        struct render_backend_frame_lifetime_allocator_page_t *next = page->next;
-        render_backend_frame_lifetime_allocator_destroy_page (system, page, destroy_buffers);
-        page = next;
-    }
-
-    kan_free_batched (system->frame_lifetime_wrapper_allocation_group, frame_lifetime_allocator);
-}
-
-static inline VkFormat kan_render_image_description_calculate_format (
-    struct render_backend_system_t *system, struct kan_render_image_description_t *description)
-{
-    VkFormat image_format = VK_FORMAT_R8G8B8A8_SRGB;
-    switch (description->type)
-    {
-    case KAN_RENDER_IMAGE_TYPE_COLOR_2D:
-    case KAN_RENDER_IMAGE_TYPE_COLOR_3D:
-        switch (description->color_format)
-        {
-        case KAN_RENDER_COLOR_FORMAT_RGBA32_SRGB:
-            image_format = VK_FORMAT_R8G8B8A8_SRGB;
-            break;
-
-        case KAN_RENDER_COLOR_FORMAT_RGBA128_SFLOAT:
-            image_format = VK_FORMAT_R32G32B32A32_SFLOAT;
-            break;
-
-        case KAN_RENDER_COLOR_FORMAT_SURFACE:
-            image_format = SURFACE_COLOR_FORMAT;
-            break;
-        }
-
-        break;
-
-    case KAN_RENDER_IMAGE_TYPE_DEPTH_STENCIL:
-        image_format = system->device_depth_image_format;
-        break;
-    }
-
-    return image_format;
-}
-
-static inline kan_bool_t create_vulkan_image (struct render_backend_system_t *system,
-                                              struct kan_render_image_description_t *description,
-                                              VkImage *output_image,
-                                              VmaAllocation *output_allocation)
-{
-    VkImageType image_type = VK_IMAGE_TYPE_2D;
-    VkImageUsageFlags image_usage = 0u;
-
-    VkSharingMode sharing_mode;
-    uint32_t sharing_index_count;
-    uint32_t shared_indices[2u] = {system->device_graphics_queue_family_index,
-                                   system->device_transfer_queue_family_index};
-
-    switch (description->type)
-    {
-    case KAN_RENDER_IMAGE_TYPE_COLOR_2D:
-        KAN_ASSERT (description->depth == 1u)
-        image_type = VK_IMAGE_TYPE_2D;
-        image_usage |= VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT;
-        break;
-
-    case KAN_RENDER_IMAGE_TYPE_COLOR_3D:
-        image_type = VK_IMAGE_TYPE_3D;
-        image_usage |= VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT;
-        break;
-
-    case KAN_RENDER_IMAGE_TYPE_DEPTH_STENCIL:
-        KAN_ASSERT (description->render_target)
-        image_type = VK_IMAGE_TYPE_2D;
-        image_usage |= VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT;
-        break;
-    }
-
-    if (description->render_target)
-    {
-        KAN_ASSERT (description->mips == 1u)
-        switch (description->type)
-        {
-        case KAN_RENDER_IMAGE_TYPE_COLOR_2D:
-        case KAN_RENDER_IMAGE_TYPE_COLOR_3D:
-            image_usage |= VK_IMAGE_USAGE_COLOR_ATTACHMENT_BIT;
-            break;
-
-        case KAN_RENDER_IMAGE_TYPE_DEPTH_STENCIL:
-            image_usage |= VK_IMAGE_USAGE_DEPTH_STENCIL_ATTACHMENT_BIT;
-            break;
-        }
-    }
-    else
-    {
-        // We need source transfer in order to generate mip maps, unless they are expected to be loaded from textures.
-        image_usage |= VK_IMAGE_USAGE_TRANSFER_SRC_BIT | VK_IMAGE_USAGE_TRANSFER_DST_BIT;
-    }
-
-    if (description->supports_sampling)
-    {
-        image_usage |= VK_IMAGE_USAGE_SAMPLED_BIT;
-    }
-
-    if (description->render_target &&
-        system->device_graphics_queue_family_index != system->device_transfer_queue_family_index)
-    {
-        sharing_mode = VK_SHARING_MODE_CONCURRENT;
-        sharing_index_count = 2u;
-    }
-    else
-    {
-        sharing_mode = VK_SHARING_MODE_EXCLUSIVE;
-        sharing_index_count = 1u;
-    }
-
-    VkImageCreateInfo image_create_info = {
-        .sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO,
-        .pNext = NULL,
-        .imageType = image_type,
-        .format = kan_render_image_description_calculate_format (system, description),
-        .extent =
-            {
-                .width = (uint32_t) description->width,
-                .height = (uint32_t) description->height,
-                .depth = (uint32_t) description->depth,
-            },
-        .mipLevels = (uint32_t) description->mips,
-        .arrayLayers = 1u,
-        .samples = VK_SAMPLE_COUNT_1_BIT,
-        .tiling = VK_IMAGE_TILING_OPTIMAL,
-        .usage = image_usage,
-        .sharingMode = sharing_mode,
-        .queueFamilyIndexCount = sharing_index_count,
-        .pQueueFamilyIndices = shared_indices,
-        .initialLayout = VK_IMAGE_LAYOUT_UNDEFINED,
-    };
-
-    VmaAllocationCreateInfo allocation_create_info = {
-        .flags = 0u,
-        .usage = VMA_MEMORY_USAGE_AUTO,
-        .requiredFlags = 0u,
-        .preferredFlags = 0u,
-        .memoryTypeBits = UINT32_MAX,
-        .pool = NULL,
-        .pUserData = NULL,
-        .priority = 0.0f,
-    };
-
-    VmaAllocationInfo allocation_info;
-    if (vmaCreateImage (system->gpu_memory_allocator, &image_create_info, &allocation_create_info, output_image,
-                        output_allocation, &allocation_info) != VK_SUCCESS)
-    {
-        KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR, "Failed to create image \"%s\".",
-                 description->tracking_name)
-        return KAN_FALSE;
-    }
-
-    return KAN_TRUE;
-}
-
-static inline VkImageAspectFlags kan_render_image_description_calculate_aspects (
-    struct render_backend_system_t *system, struct kan_render_image_description_t *description)
-{
-    VkImageAspectFlags aspects = 0u;
-    switch (description->type)
-    {
-    case KAN_RENDER_IMAGE_TYPE_COLOR_2D:
-    case KAN_RENDER_IMAGE_TYPE_COLOR_3D:
-        aspects |= VK_IMAGE_ASPECT_COLOR_BIT;
-        break;
-
-    case KAN_RENDER_IMAGE_TYPE_DEPTH_STENCIL:
-        aspects |= VK_IMAGE_ASPECT_DEPTH_BIT;
-        if (system->device_depth_image_has_stencil)
-        {
-            aspects |= VK_IMAGE_ASPECT_STENCIL_BIT;
-        }
-
-        break;
-    }
-
-    return aspects;
-}
-
-static inline void kan_render_image_description_calculate_size_at_mip (
-    struct kan_render_image_description_t *description,
-    uint64_t mip,
-    uint64_t *output_width,
-    uint64_t *output_height,
-    uint64_t *output_depth)
-{
-    KAN_ASSERT (mip < description->mips)
-    *output_width = KAN_MAX (1u, description->width >> mip);
-    *output_height = KAN_MAX (1u, description->height >> mip);
-    *output_depth = KAN_MAX (1u, description->depth >> mip);
-}
-
-static inline uint64_t kan_render_image_description_calculate_texel_size (
-    struct render_backend_system_t *system, struct kan_render_image_description_t *description)
-{
-    uint64_t texel_size = 0u;
-    switch (description->type)
-    {
-    case KAN_RENDER_IMAGE_TYPE_COLOR_2D:
-    case KAN_RENDER_IMAGE_TYPE_COLOR_3D:
-        switch (description->color_format)
-        {
-        case KAN_RENDER_COLOR_FORMAT_RGBA32_SRGB:
-        case KAN_RENDER_COLOR_FORMAT_SURFACE:
-            texel_size = 4u;
-            break;
-
-        case KAN_RENDER_COLOR_FORMAT_RGBA128_SFLOAT:
-            texel_size = 16u;
-            break;
-        }
-
-        break;
-
-    case KAN_RENDER_IMAGE_TYPE_DEPTH_STENCIL:
-        switch (system->device_depth_image_format)
-        {
-        case VK_FORMAT_D32_SFLOAT:
-        case VK_FORMAT_D24_UNORM_S8_UINT:
-            texel_size = 4u;
-            break;
-
-        case VK_FORMAT_D32_SFLOAT_S8_UINT:
-            texel_size = 5u;
-            break;
-
-        default:
-            KAN_ASSERT (KAN_FALSE)
-            texel_size = 0u;
-            break;
-        }
-
-        break;
-    }
-
-    return texel_size;
-}
-
-#if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_PROFILE_MEMORY)
-static uint64_t render_backend_image_calculate_gpu_size (struct render_backend_system_t *system,
-                                                         struct render_backend_image_t *image)
-{
-    uint64_t texel_size = kan_render_image_description_calculate_texel_size (system, &image->description);
-    uint64_t size = 0u;
-
-    for (uint64_t mip = 0u; mip < image->description.mips; ++mip)
-    {
-        uint64_t width;
-        uint64_t height;
-        uint64_t depth;
-        kan_render_image_description_calculate_size_at_mip (&image->description, mip, &width, &height, &depth);
-        size += texel_size * width * height * depth;
-    }
-
-    return size;
-}
-#endif
-
-static struct render_backend_image_t *render_backend_system_create_image (
-    struct render_backend_system_t *system, struct kan_render_image_description_t *description)
-{
-    VkImage vulkan_image;
-    VmaAllocation vulkan_allocation;
-
-    if (!create_vulkan_image (system, description, &vulkan_image, &vulkan_allocation))
-    {
-        return NULL;
-    }
-
-    struct render_backend_image_t *image =
-        kan_allocate_batched (system->image_wrapper_allocation_group, sizeof (struct render_backend_image_t));
-    image->next = system->first_image;
-    image->previous = NULL;
-
-    if (system->first_image)
-    {
-        system->first_image->previous = image;
-    }
-
-    system->first_image = image;
-    image->system = system;
-
-    image->image = vulkan_image;
-    image->allocation = vulkan_allocation;
-    image->description = *description;
-    image->switched_to_transfer_source = KAN_FALSE;
-    image->first_frame_buffer_attachment = NULL;
-
-#if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_PROFILE_MEMORY)
-    image->device_allocation_group =
-        kan_allocation_group_get_child (system->memory_profiling.gpu_image_group, description->tracking_name);
-    transfer_memory_between_groups (render_backend_image_calculate_gpu_size (system, image),
-                                    system->memory_profiling.gpu_unmarked_group, image->device_allocation_group);
-#endif
-
-    return image;
-}
-
-static void render_backend_system_destroy_image (struct render_backend_system_t *system,
-                                                 struct render_backend_image_t *image,
-                                                 kan_bool_t remove_from_list)
-{
-    if (remove_from_list)
-    {
-        if (image->next)
-        {
-            image->next->previous = image->previous;
-        }
-
-        if (image->previous)
-        {
-            image->previous->next = image->next;
-        }
-        else
-        {
-            KAN_ASSERT (system->first_image == image)
-            system->first_image = image->next;
-        }
-    }
-
-    struct image_frame_buffer_attachment_t *attachment = image->first_frame_buffer_attachment;
-    while (attachment)
-    {
-        struct image_frame_buffer_attachment_t *next = attachment->next;
-        kan_free_batched (system->image_wrapper_allocation_group, attachment);
-        attachment = next;
-    }
-
-#if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_PROFILE_MEMORY)
-    transfer_memory_between_groups (render_backend_image_calculate_gpu_size (system, image),
-                                    image->device_allocation_group, system->memory_profiling.gpu_unmarked_group);
-#endif
-
-    vmaDestroyImage (system->gpu_memory_allocator, image->image, image->allocation);
-    kan_free_batched (system->image_wrapper_allocation_group, image);
-}
 
 kan_context_system_handle_t render_backend_system_create (kan_allocation_group_t group, void *user_config)
 {
@@ -2711,7 +265,7 @@ void render_backend_system_init (kan_context_system_handle_t handle)
 #endif
 
 #if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_PROFILE_MEMORY)
-    memory_profiling_init (system);
+    render_backend_memory_profiling_init (system);
 #endif
 
     VkApplicationInfo application_info = {
@@ -3022,7 +576,7 @@ void render_backend_system_destroy (kan_context_system_handle_t handle)
 }
 
 struct kan_context_system_api_t KAN_CONTEXT_SYSTEM_API_NAME (render_backend_system_t) = {
-    .name = "render_backend_system_t",
+    .name = KAN_CONTEXT_RENDER_BACKEND_SYSTEM_NAME,
     .create = render_backend_system_create,
     .connect = render_backend_system_connect,
     .connected_init = render_backend_system_init,
@@ -5313,560 +2867,4 @@ void kan_render_backend_system_destroy_surface (kan_context_system_handle_t rend
 enum kan_platform_window_flag_t kan_render_get_required_window_flags (void)
 {
     return KAN_PLATFORM_WINDOW_FLAG_SUPPORTS_VULKAN;
-}
-
-kan_render_frame_buffer_t kan_render_frame_buffer_create (kan_render_context_t context,
-                                                          struct kan_render_frame_buffer_description_t *description)
-{
-    struct render_backend_system_t *system = (struct render_backend_system_t *) context;
-    kan_atomic_int_lock (&system->resource_management_lock);
-    struct render_backend_frame_buffer_t *frame_buffer =
-        render_backend_system_create_frame_buffer (system, description);
-    kan_atomic_int_unlock (&system->resource_management_lock);
-    return frame_buffer ? (kan_render_frame_buffer_t) frame_buffer : KAN_INVALID_FRAME_BUFFER;
-}
-
-void kan_render_frame_buffer_destroy (kan_render_frame_buffer_t buffer)
-{
-    struct render_backend_frame_buffer_t *data = (struct render_backend_frame_buffer_t *) buffer;
-    struct render_backend_schedule_state_t *schedule = render_backend_system_get_schedule_for_destroy (data->system);
-    kan_atomic_int_lock (&schedule->schedule_lock);
-
-    struct scheduled_frame_buffer_destroy_t *item =
-        KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (&schedule->item_allocator, struct scheduled_frame_buffer_destroy_t);
-
-    // We do not need to preserve order as frame buffers cannot depend one on another.
-    item->next = schedule->first_scheduled_frame_buffer_destroy;
-    schedule->first_scheduled_frame_buffer_destroy = item;
-    item->frame_buffer = data;
-    kan_atomic_int_unlock (&schedule->schedule_lock);
-}
-
-kan_render_pass_t kan_render_pass_create (kan_render_context_t context,
-                                          struct kan_render_pass_description_t *description)
-{
-    struct render_backend_system_t *system = (struct render_backend_system_t *) context;
-    kan_atomic_int_lock (&system->resource_management_lock);
-    struct render_backend_pass_t *pass = render_backend_system_create_pass (system, description);
-    kan_atomic_int_unlock (&system->resource_management_lock);
-    return pass ? (kan_render_pass_t) pass : KAN_INVALID_RENDER_PASS;
-}
-
-kan_bool_t kan_render_pass_add_static_dependency (kan_render_pass_t pass, kan_render_pass_t dependency)
-{
-    // TODO: Implement.
-    return KAN_FALSE;
-}
-
-kan_render_pass_instance_t kan_render_pass_instantiate (kan_render_pass_t pass,
-                                                        kan_render_frame_buffer_t frame_buffer,
-                                                        struct kan_render_viewport_bounds_t *viewport_bounds,
-                                                        struct kan_render_integer_region_t *scissor)
-{
-    // TODO: Implement.
-    return KAN_INVALID_RENDER_PASS_INSTANCE;
-}
-
-void kan_render_pass_instance_add_dynamic_dependency (kan_render_pass_instance_t pass_instance,
-                                                      kan_render_pass_instance_t dependency)
-{
-    // TODO: Implement.
-}
-
-void kan_render_pass_instance_classic_graphics_pipeline (
-    kan_render_pass_instance_t pass_instance, kan_render_classic_graphics_pipeline_instance_t pipeline_instance)
-{
-    // TODO: Implement.
-}
-
-void kan_render_pass_instance_attributes (kan_render_pass_instance_t pass_instance,
-                                          uint64_t start_at_binding,
-                                          uint64_t buffers_count,
-                                          kan_render_buffer_t *buffers)
-{
-    // TODO: Implement.
-}
-
-void kan_render_pass_instance_indices (kan_render_pass_instance_t pass_instance, kan_render_buffer_t buffer)
-{
-    // TODO: Implement.
-}
-
-void kan_render_pass_instance_draw (kan_render_pass_instance_t pass_instance,
-                                    uint64_t vertex_offset,
-                                    uint64_t vertex_count)
-{
-    // TODO: Implement.
-}
-
-void kan_render_pass_instance_instanced_draw (kan_render_pass_instance_t pass_instance,
-                                              uint64_t vertex_offset,
-                                              uint64_t vertex_count,
-                                              uint64_t instance_offset,
-                                              uint64_t instance_count)
-{
-    // TODO: Implement.
-}
-
-void kan_render_pass_destroy (kan_render_pass_t pass)
-{
-    struct render_backend_pass_t *data = (struct render_backend_pass_t *) pass;
-    struct render_backend_schedule_state_t *schedule = render_backend_system_get_schedule_for_destroy (data->system);
-    kan_atomic_int_lock (&schedule->schedule_lock);
-
-    struct scheduled_pass_destroy_t *item =
-        KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (&schedule->item_allocator, struct scheduled_pass_destroy_t);
-
-    // We do not need to preserve order as passes cannot depend one on another.
-    item->next = schedule->first_scheduled_pass_destroy;
-    schedule->first_scheduled_pass_destroy = item;
-    item->pass = data;
-    kan_atomic_int_unlock (&schedule->schedule_lock);
-}
-
-kan_render_classic_graphics_pipeline_family_t kan_render_classic_graphics_pipeline_family_create (
-    kan_render_context_t context, struct kan_render_classic_graphics_pipeline_family_description_t *description)
-{
-    struct render_backend_system_t *system = (struct render_backend_system_t *) context;
-    kan_atomic_int_lock (&system->resource_management_lock);
-    struct render_backend_classic_graphics_pipeline_family_t *family =
-        render_backend_system_create_classic_graphics_pipeline_family (system, description);
-    kan_atomic_int_unlock (&system->resource_management_lock);
-    return family ? (kan_render_classic_graphics_pipeline_family_t) family :
-                    KAN_INVALID_RENDER_CLASSIC_GRAPHICS_PIPELINE_FAMILY;
-}
-
-void kan_render_classic_graphics_pipeline_family_destroy (kan_render_classic_graphics_pipeline_family_t family)
-{
-    struct render_backend_classic_graphics_pipeline_family_t *data =
-        (struct render_backend_classic_graphics_pipeline_family_t *) family;
-    struct render_backend_schedule_state_t *schedule = render_backend_system_get_schedule_for_destroy (data->system);
-    kan_atomic_int_lock (&schedule->schedule_lock);
-
-    struct scheduled_classic_graphics_pipeline_family_destroy_t *item = KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (
-        &schedule->item_allocator, struct scheduled_classic_graphics_pipeline_family_destroy_t);
-
-    item->next = schedule->first_scheduled_classic_graphics_pipeline_family_destroy;
-    schedule->first_scheduled_classic_graphics_pipeline_family_destroy = item;
-    item->family = data;
-    kan_atomic_int_unlock (&schedule->schedule_lock);
-}
-
-kan_render_classic_graphics_pipeline_t kan_render_classic_graphics_pipeline_create (
-    kan_render_context_t context,
-    struct kan_render_classic_graphics_pipeline_description_t *description,
-    enum kan_render_pipeline_compilation_priority_t compilation_priority)
-{
-    // TODO: Implement.
-    return KAN_INVALID_RENDER_CLASSIC_GRAPHICS_PIPELINE;
-}
-
-void kan_render_classic_graphics_pipeline_change_compilation_priority (
-    kan_render_classic_graphics_pipeline_t pipeline,
-    enum kan_render_pipeline_compilation_priority_t compilation_priority)
-{
-    // TODO: Implement.
-}
-
-CONTEXT_RENDER_BACKEND_SYSTEM_API void kan_render_classic_graphics_pipeline_destroy (
-    kan_render_classic_graphics_pipeline_t pipeline)
-{
-    // TODO: Implement.
-}
-
-kan_render_classic_graphics_pipeline_instance_t kan_render_classic_graphics_pipeline_instance_create (
-    kan_render_context_t context,
-    kan_render_classic_graphics_pipeline_t pipeline,
-    uint64_t initial_bindings_count,
-    struct kan_render_layout_update_description_t *initial_bindings,
-    kan_interned_string_t tracking_name)
-{
-    // TODO: Implement.
-    return KAN_INVALID_RENDER_CLASSIC_GRAPHICS_PIPELINE_INSTANCE;
-}
-
-void kan_render_classic_graphics_pipeline_instance_update_layout (
-    kan_render_classic_graphics_pipeline_instance_t instance,
-    uint64_t bindings_count,
-    struct kan_render_layout_update_description_t *bindings)
-{
-    // TODO: Implement.
-}
-
-CONTEXT_RENDER_BACKEND_SYSTEM_API void kan_render_classic_graphics_pipeline_instance_destroy (
-    kan_render_classic_graphics_pipeline_instance_t instance)
-{
-    // TODO: Implement.
-}
-
-kan_render_buffer_t kan_render_buffer_create (kan_render_context_t context,
-                                              enum kan_render_buffer_type_t type,
-                                              uint64_t full_size,
-                                              void *optional_initial_data,
-                                              kan_interned_string_t tracking_name)
-{
-    struct render_backend_system_t *system = (struct render_backend_system_t *) context;
-    kan_atomic_int_lock (&system->resource_management_lock);
-    struct render_backend_buffer_t *buffer = render_backend_system_create_buffer (
-        system, RENDER_BACKEND_BUFFER_FAMILY_RESOURCE, type, full_size, tracking_name);
-    kan_atomic_int_unlock (&system->resource_management_lock);
-
-    if (!buffer)
-    {
-        return KAN_INVALID_RENDER_BUFFER;
-    }
-
-    if (optional_initial_data)
-    {
-        if (system->device_memory_type == KAN_RENDER_DEVICE_MEMORY_TYPE_UNIFIED ||
-            system->device_memory_type == KAN_RENDER_DEVICE_MEMORY_TYPE_UNIFIED_COHERENT)
-        {
-            // Due to unified memory, we should be able to directly initialize buffer.
-            void *memory;
-            kan_atomic_int_lock (&system->resource_management_lock);
-
-            if (vmaMapMemory (system->gpu_memory_allocator, buffer->allocation, &memory) != VK_SUCCESS)
-            {
-                kan_critical_error ("Unexpected failure while mapping buffer memory, unable to continue properly.",
-                                    __FILE__, __LINE__);
-            }
-
-            kan_atomic_int_unlock (&system->resource_management_lock);
-            memcpy (memory, optional_initial_data, full_size);
-            kan_atomic_int_lock (&system->resource_management_lock);
-            vmaUnmapMemory (system->gpu_memory_allocator, buffer->allocation);
-
-            if (vmaFlushAllocation (system->gpu_memory_allocator, buffer->allocation, 0u, full_size) != VK_SUCCESS)
-            {
-                kan_critical_error (
-                    "Unexpected failure while flushing buffer initial data, unable to continue properly.", __FILE__,
-                    __LINE__);
-            }
-
-            kan_atomic_int_unlock (&system->resource_management_lock);
-        }
-        else
-        {
-            void *data = kan_render_buffer_patch ((kan_render_buffer_t) buffer, 0u, full_size);
-            // Patch can fail if there is no available staging memory,
-            // but we have no special handling for that right now.
-            KAN_ASSERT (data)
-            memcpy (data, optional_initial_data, full_size);
-        }
-    }
-
-    return (kan_render_buffer_t) buffer;
-}
-
-void *kan_render_buffer_patch (kan_render_buffer_t buffer, uint64_t slice_offset, uint64_t slice_size)
-{
-    struct render_backend_buffer_t *data = (struct render_backend_buffer_t *) buffer;
-    struct render_backend_schedule_state_t *schedule = render_backend_system_get_schedule_for_memory (data->system);
-
-    switch (data->family)
-    {
-    case RENDER_BACKEND_BUFFER_FAMILY_RESOURCE:
-    {
-        struct render_backend_frame_lifetime_allocator_allocation_t staging_allocation =
-            render_backend_system_allocate_for_staging (data->system, slice_size);
-
-        if (!staging_allocation.buffer)
-        {
-            KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR,
-                     "Failed to patch buffer \"%s\": out of staging memory.", data->tracking_name)
-            return NULL;
-        }
-
-        kan_atomic_int_lock (&data->system->resource_management_lock);
-        void *memory;
-
-        if (vmaMapMemory (data->system->gpu_memory_allocator, staging_allocation.buffer->allocation, &memory) !=
-            VK_SUCCESS)
-        {
-            kan_critical_error ("Unexpected failure while mapping staging buffer memory, unable to continue properly.",
-                                __FILE__, __LINE__);
-        }
-
-        kan_atomic_int_unlock (&data->system->resource_management_lock);
-        kan_atomic_int_lock (&schedule->schedule_lock);
-
-        struct scheduled_buffer_unmap_flush_transfer_t *item = KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (
-            &schedule->item_allocator, struct scheduled_buffer_unmap_flush_transfer_t);
-
-        // We do not need to preserve order as buffers cannot depend one on another.
-        item->next = schedule->first_scheduled_buffer_unmap_flush_transfer;
-        schedule->first_scheduled_buffer_unmap_flush_transfer = item;
-        item->source_buffer = staging_allocation.buffer;
-        item->target_buffer = data;
-        item->source_offset = staging_allocation.offset;
-        item->target_offset = slice_offset;
-        item->size = slice_size;
-        kan_atomic_int_unlock (&schedule->schedule_lock);
-
-        return ((uint8_t *) memory) + staging_allocation.offset;
-    }
-
-    case RENDER_BACKEND_BUFFER_FAMILY_STAGING:
-        // Staging buffers are not exposed to user, we shouldn't be able to get here normally.
-        KAN_ASSERT (KAN_FALSE)
-        break;
-
-    case RENDER_BACKEND_BUFFER_FAMILY_FRAME_LIFETIME_ALLOCATOR:
-    {
-        // Frame lifetime allocations are always host visible, there is no need to stage them due to their lifetime.
-        void *memory;
-        kan_atomic_int_lock (&data->system->resource_management_lock);
-
-        if (vmaMapMemory (data->system->gpu_memory_allocator, data->allocation, &memory) != VK_SUCCESS)
-        {
-            kan_critical_error ("Unexpected failure while mapping buffer memory, unable to continue properly.",
-                                __FILE__, __LINE__);
-        }
-
-        kan_atomic_int_unlock (&data->system->resource_management_lock);
-        kan_atomic_int_lock (&schedule->schedule_lock);
-
-        struct scheduled_buffer_unmap_flush_t *item =
-            KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (&schedule->item_allocator, struct scheduled_buffer_unmap_flush_t);
-
-        // We do not need to preserve order as buffers cannot depend one on another.
-        item->next = schedule->first_scheduled_buffer_unmap_flush;
-        schedule->first_scheduled_buffer_unmap_flush = item;
-        item->buffer = data;
-        item->offset = slice_offset;
-        item->size = slice_size;
-        kan_atomic_int_unlock (&schedule->schedule_lock);
-        return ((uint8_t *) memory) + slice_offset;
-    }
-    }
-
-    KAN_ASSERT (KAN_FALSE)
-    return NULL;
-}
-
-void kan_render_buffer_destroy (kan_render_buffer_t buffer)
-{
-    struct render_backend_buffer_t *data = (struct render_backend_buffer_t *) buffer;
-    // Only resource family buffers can be destroyed externally through scheduling.
-    KAN_ASSERT (data->family == RENDER_BACKEND_BUFFER_FAMILY_RESOURCE)
-
-    struct render_backend_schedule_state_t *schedule = render_backend_system_get_schedule_for_destroy (data->system);
-    kan_atomic_int_lock (&schedule->schedule_lock);
-
-    struct scheduled_buffer_destroy_t *item =
-        KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (&schedule->item_allocator, struct scheduled_buffer_destroy_t);
-
-    // We do not need to preserve order as buffers cannot depend one on another.
-    item->next = schedule->first_scheduled_buffer_destroy;
-    schedule->first_scheduled_buffer_destroy = item;
-    item->buffer = data;
-    kan_atomic_int_unlock (&schedule->schedule_lock);
-}
-
-kan_render_frame_lifetime_buffer_allocator_t kan_render_frame_lifetime_buffer_allocator_create (
-    kan_render_context_t context,
-    enum kan_render_buffer_type_t buffer_type,
-    uint64_t page_size,
-    kan_interned_string_t tracking_name)
-{
-    struct render_backend_system_t *system = (struct render_backend_system_t *) context;
-    kan_atomic_int_lock (&system->resource_management_lock);
-    struct render_backend_frame_lifetime_allocator_t *allocator =
-        render_backend_system_create_frame_lifetime_allocator (
-            system, RENDER_BACKEND_BUFFER_FAMILY_FRAME_LIFETIME_ALLOCATOR, buffer_type, page_size, tracking_name);
-    kan_atomic_int_unlock (&system->resource_management_lock);
-    return (kan_render_frame_lifetime_buffer_allocator_t) allocator;
-}
-
-struct kan_render_allocated_slice_t kan_render_frame_lifetime_buffer_allocator_allocate (
-    kan_render_frame_lifetime_buffer_allocator_t allocator, uint64_t size, uint64_t alignment)
-{
-    struct render_backend_frame_lifetime_allocator_t *data =
-        (struct render_backend_frame_lifetime_allocator_t *) allocator;
-
-    struct render_backend_frame_lifetime_allocator_allocation_t allocation =
-        render_backend_frame_lifetime_allocator_allocate (data, size, alignment);
-
-    return (struct kan_render_allocated_slice_t) {
-        .buffer = (kan_render_buffer_t) allocation.buffer,
-        .slice_offset = allocation.offset,
-    };
-}
-
-void kan_render_frame_lifetime_buffer_allocator_destroy (kan_render_frame_lifetime_buffer_allocator_t allocator)
-{
-    struct render_backend_frame_lifetime_allocator_t *data =
-        (struct render_backend_frame_lifetime_allocator_t *) allocator;
-    // Only resource family buffers can be destroyed externally through scheduling.
-    KAN_ASSERT (data->buffer_family == RENDER_BACKEND_BUFFER_FAMILY_FRAME_LIFETIME_ALLOCATOR)
-
-    struct render_backend_schedule_state_t *schedule = render_backend_system_get_schedule_for_destroy (data->system);
-    kan_atomic_int_lock (&schedule->schedule_lock);
-
-    struct scheduled_frame_lifetime_allocator_destroy_t *item = KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (
-        &schedule->item_allocator, struct scheduled_frame_lifetime_allocator_destroy_t);
-
-    // We do not need to preserve order as frame_lifetime_allocators cannot depend one on another.
-    item->next = schedule->first_scheduled_frame_lifetime_allocator_destroy;
-    schedule->first_scheduled_frame_lifetime_allocator_destroy = item;
-    item->frame_lifetime_allocator = data;
-    kan_atomic_int_unlock (&schedule->schedule_lock);
-}
-
-kan_render_image_t kan_render_image_create (kan_render_context_t context,
-                                            struct kan_render_image_description_t *description)
-{
-    struct render_backend_system_t *system = (struct render_backend_system_t *) context;
-    kan_atomic_int_lock (&system->resource_management_lock);
-    struct render_backend_image_t *image = render_backend_system_create_image (system, description);
-    kan_atomic_int_unlock (&system->resource_management_lock);
-    return image ? (kan_render_image_t) image : KAN_INVALID_RENDER_IMAGE;
-}
-
-void kan_render_image_upload_data (kan_render_image_t image, uint64_t mip, void *data)
-{
-    struct render_backend_image_t *image_data = (struct render_backend_image_t *) image;
-    KAN_ASSERT (!image_data->description.render_target)
-    KAN_ASSERT (mip < image_data->description.mips)
-
-    uint64_t texel_size =
-        kan_render_image_description_calculate_texel_size (image_data->system, &image_data->description);
-
-    uint64_t width;
-    uint64_t height;
-    uint64_t depth;
-    kan_render_image_description_calculate_size_at_mip (&image_data->description, mip, &width, &height, &depth);
-
-    const uint64_t allocation_size = texel_size * width * height * depth;
-    struct render_backend_frame_lifetime_allocator_allocation_t staging_allocation =
-        render_backend_system_allocate_for_staging (image_data->system, allocation_size);
-
-    if (!staging_allocation.buffer)
-    {
-        KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR,
-                 "Failed to upload image \"%s\" mip %lu: out of staging memory.", image_data->description.tracking_name,
-                 (unsigned long) mip)
-        return;
-    }
-
-    struct render_backend_schedule_state_t *schedule =
-        render_backend_system_get_schedule_for_memory (image_data->system);
-    kan_atomic_int_lock (&schedule->schedule_lock);
-
-    struct scheduled_image_upload_t *item =
-        KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (&schedule->item_allocator, struct scheduled_image_upload_t);
-
-    // We do not need to preserve order as images cannot depend one on another.
-    item->next = schedule->first_scheduled_image_upload;
-    schedule->first_scheduled_image_upload = item;
-    item->image = image_data;
-    item->mip = mip;
-    item->staging_buffer = staging_allocation.buffer;
-    item->staging_buffer_offset = staging_allocation.offset;
-    kan_atomic_int_unlock (&schedule->schedule_lock);
-}
-
-void kan_render_image_request_mip_generation (kan_render_image_t image, uint64_t first, uint64_t last)
-{
-    struct render_backend_image_t *data = (struct render_backend_image_t *) image;
-    KAN_ASSERT (!data->description.render_target)
-
-    struct render_backend_schedule_state_t *schedule = render_backend_system_get_schedule_for_memory (data->system);
-    kan_atomic_int_lock (&schedule->schedule_lock);
-
-    struct scheduled_image_mip_generation_t *item =
-        KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (&schedule->item_allocator, struct scheduled_image_mip_generation_t);
-
-    // We do not need to preserve order as images cannot depend one on another.
-    item->next = schedule->first_scheduled_image_mip_generation;
-    schedule->first_scheduled_image_mip_generation = item;
-    item->image = data;
-    item->first = first;
-    item->last = last;
-    kan_atomic_int_unlock (&schedule->schedule_lock);
-}
-
-void kan_render_image_resize_render_target (kan_render_image_t image,
-                                            uint64_t new_width,
-                                            uint64_t new_height,
-                                            uint64_t new_depth)
-{
-    struct render_backend_image_t *data = (struct render_backend_image_t *) image;
-    KAN_ASSERT (data->description.render_target)
-
-    // If it is null handle, then we've tried to resize earlier and failed. No need for additional cleanup.
-    if (data->image != VK_NULL_HANDLE)
-    {
-        struct render_backend_schedule_state_t *schedule =
-            render_backend_system_get_schedule_for_destroy (data->system);
-        kan_atomic_int_lock (&schedule->schedule_lock);
-
-        struct image_frame_buffer_attachment_t *attachment = data->first_frame_buffer_attachment;
-        while (attachment)
-        {
-            render_backend_frame_buffer_schedule_resource_destroy (data->system, attachment->frame_buffer, schedule);
-            attachment = attachment->next;
-        }
-
-        struct scheduled_detached_image_destroy_t *image_destroy = KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (
-            &schedule->item_allocator, struct scheduled_detached_image_destroy_t);
-
-        image_destroy->next = schedule->first_scheduled_detached_image_destroy;
-        schedule->first_scheduled_detached_image_destroy = image_destroy;
-        image_destroy->detached_image = data->image;
-        image_destroy->detached_allocation = data->allocation;
-
-#if defined(KAN_CONTEXT_RENDER_BACKEND_VULKAN_PROFILE_MEMORY)
-        image_destroy->gpu_size = render_backend_image_calculate_gpu_size (data->system, data);
-        image_destroy->gpu_allocation_group = data->device_allocation_group;
-#endif
-
-        kan_atomic_int_unlock (&schedule->schedule_lock);
-    }
-
-    data->description.width = new_width;
-    data->description.height = new_height;
-    data->description.depth = new_depth;
-
-    if (!create_vulkan_image (data->system, &data->description, &data->image, &data->allocation))
-    {
-        data->image = VK_NULL_HANDLE;
-        KAN_LOG (render_backend_system_vulkan, KAN_LOG_ERROR,
-                 "Failed to resize image \"%s\": new image creation failed.", data->description.tracking_name)
-        return;
-    }
-
-    struct render_backend_schedule_state_t *schedule = render_backend_system_get_schedule_for_memory (data->system);
-    kan_atomic_int_lock (&schedule->schedule_lock);
-    struct image_frame_buffer_attachment_t *attachment = data->first_frame_buffer_attachment;
-
-    while (attachment)
-    {
-        struct scheduled_frame_buffer_create_t *item = KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (
-            &schedule->item_allocator, struct scheduled_frame_buffer_create_t);
-
-        // We do not need to preserve order as frame buffers cannot depend one on another.
-        item->next = schedule->first_scheduled_frame_buffer_create;
-        schedule->first_scheduled_frame_buffer_create = item;
-        item->frame_buffer = attachment->frame_buffer;
-        attachment = attachment->next;
-    }
-
-    kan_atomic_int_unlock (&schedule->schedule_lock);
-}
-
-void kan_render_image_destroy (kan_render_image_t image)
-{
-    struct render_backend_image_t *data = (struct render_backend_image_t *) image;
-    struct render_backend_schedule_state_t *schedule = render_backend_system_get_schedule_for_destroy (data->system);
-    kan_atomic_int_lock (&schedule->schedule_lock);
-
-    struct scheduled_image_destroy_t *item =
-        KAN_STACK_GROUP_ALLOCATOR_ALLOCATE_TYPED (&schedule->item_allocator, struct scheduled_image_destroy_t);
-
-    // We do not need to preserve order as images cannot depend one on another.
-    item->next = schedule->first_scheduled_image_destroy;
-    schedule->first_scheduled_image_destroy = item;
-    item->image = data;
-    kan_atomic_int_unlock (&schedule->schedule_lock);
 }
